% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc2sp.R
\name{arc2sp}
\alias{arc2sp}
\title{Converts ArcGIS feature class to an \code{sp} object}
\usage{
arc2sp(fc_path)
}
\arguments{
\item{fc_path}{string; Path to the ArcGIS feature class.}
}
\value{
The specified ArcGIS feature class as an \code{sp} object.
}
\description{
Opens an ArcGIS feature class and converts it to the \code{sp}
    spatial data format.
}
