% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bend_raduis.R
\name{bend_radius}
\alias{bend_radius}
\title{Calculate the bend radius of curvature}
\usage{
bend_radius(bankline_points)
}
\arguments{
\item{bankline_points}{sf data frame; an fluvgeo bankline_points
data structure}
}
\value{
Returns a data frame of bends with the calculated raduis of
curvature.
}
\description{
Calculates the radius of curvature for each bend in the input
`bankline_points` data frame.
}
