% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bend_raduis_plot.R
\name{bend_raduis_plot}
\alias{bend_raduis_plot}
\title{Plot the bend radius of curvature}
\usage{
bend_raduis_plot(bankline_points, loop, bend, coord_system)
}
\arguments{
\item{bankline_points}{sf data frame; an fluvgeo `bankline_points`
data structure}

\item{loop}{numeric; the loop to plot}

\item{bend}{numeric; the bend to plot}

\item{coord_system}{character; a brief text description of the coordinate
system. This gets added to the latitude and longitude
as an indicator of units.}
}
\value{
a ggplot2 object
}
\description{
Produces a plot of the raduis of curvature for the specified
loop and bend using a `bankline_points` data frame.
}
