% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gof_stats.R
\name{build_gof_stats}
\alias{build_gof_stats}
\title{Calculate goodness-of-fit statistics}
\usage{
build_gof_stats(xs_dims, streams, regions, bankfull_elevations)
}
\arguments{
\item{xs_dims}{data frame; a data frame of cross section
dimensions}

\item{streams}{character vector; The stream names in the study
area.}

\item{regions}{character; The regions that a dimension will be
calculated for. See the regional_curves$region
field for a complete list.}

\item{bankfull_elevations}{numeric vector; The bankfull elevations (units:
detrended feet) that are used to calculate
hydraulic geometry.}
}
\value{
A data frame of goodness of fit scores  of the relationship
   between cross section geometry and regional hydraulic geometry
   dimensions.
   \describe{
       \item{reach_name}{character; The name of the stream.}
       \item{region}{character; The name of the region.}
       \item{bankfull_elevation}{numeric; The detrended bankfull elevation
                      (in feet) that is used to calculate hydraulic
                      geometry.}
       \item{rmse_area}{numeric; The root-mean-square-error between the
                        regional curve derived area and the cross section
                        derived area.}
       \item{rmse_width}{numeric; The root-mean-square-error between the
                         regional curve derived width and the cross section
                         derived width.}
       \item{rmse_depth}{numeric; The root-mean-square-error between the
                         regional curve derived depth and the cross section
                         derived depth.}
       \item{mae_area}{numeric; The mean absolute error between the
                       regional curve derived area and the cross section
                       derived area.}
       \item{mae_width}{numeric; The mean absolute error between the
                        regional curve derived width and the cross section
                        derived width.}
       \item{mae_depth}{numeric; The mean absolute error between the
                        regional curve derived depth and the cross section
                        derived depth.}
   }
}
\description{
Calculates goodness-of-fit statistics beteween cross section
    geometry and regional hydraulic geometry dimensions.
}
