% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bankline_points.R
\name{check_bankline_points}
\alias{check_bankline_points}
\title{Check the validity of an `fgm` `bankline_points` data structure}
\usage{
check_bankline_points(bankline_points)
}
\arguments{
\item{bankline_points}{SpatialPointsDataFrame; a `bankline_points` data
structure used by the fgm package.}
}
\value{
Returns TRUE if the `bankline_points` data structure matches the
requirements. The function throws an error for a data structure not matching
the data specification. Returns errors describing how the the data structure
doesn't match the requirement.
}
\description{
Checks that the input data structure `bankline_points` meets
the requirements for this data structure.
}
