% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_channel_feature.R
\name{check_channel_feature}
\alias{check_channel_feature}
\title{Check the validity of an `fgm` `channel_feature` data structure}
\usage{
check_channel_feature(channel_feature)
}
\arguments{
\item{channel_feature}{data frame; a `channel_feature` data structure used
by the fgm package.}
}
\value{
Returns TRUE if the `channel_feature` data structure matches the
requirements. The function throws an error for a data structure not matching
the data specification. Returns errors describing how the the data structure
doesn't match the requirement.
}
\description{
Checks that the input data structure `channel_feature` meets
the requirements for this data structure.
}
