% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_section_points.R
\name{check_cross_section_points}
\alias{check_cross_section_points}
\title{Check the validity of an `fluvgeo` `cross_section_points` data structure}
\usage{
check_cross_section_points(xs_points, step = c("station_points", "loop_bend"))
}
\arguments{
\item{xs_points}{sf; a `cross_section_points`
data structure used by the fluvgeo package.}

\item{step}{character; last completed processing step. One of
"station_points", "loop_bend"}
}
\value{
Returns TRUE if the `cross_section_points` data structure matches the
requirements. The function throws an error for a data structure not matching
the data specification. Returns errors describing how the the data structure
doesn't match the requirement.
}
\description{
Checks that the input data structure `cross_section_points`
meets the requirements for this data structure.
}
