% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crs_3857.R
\name{check_crs_3857}
\alias{check_crs_3857}
\title{Test if the CRS is 3857}
\usage{
check_crs_3857(x)
}
\arguments{
\item{x}{sf, terra}
}
\value{
logical
}
\description{
Check if the input object has CRS EPSG:3857.
}
