% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_features.R
\name{check_features}
\alias{check_features}
\title{Check the validity of an `fgm` `features` data structure}
\usage{
check_features(features)
}
\arguments{
\item{features}{SpatialPointsDataFrame; a `features` data structure used
by the fgm package.}
}
\value{
Returns TRUE if the `features` data structure matches the
requirements. The function throws an error for a data structure not matching
the data specification. Returns errors describing how the the data structure
doesn't match the requirement.
}
\description{
Checks that the input data structure `features` meets
the requirements for this data structure.
}
