% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_loop_points.R
\name{check_loop_points}
\alias{check_loop_points}
\title{Check the validity of an `fluvgeo` `loop_points` data structure}
\usage{
check_loop_points(loop_points)
}
\arguments{
\item{loop_points}{sf object; a `loop_points` data structure
used by the fluvgeo package.}
}
\value{
Returns TRUE if the `loop_points` data structure matches the
requirements. The function throws an error for a data structure not matching
the data specification. Returns errors describing how the the data structure
doesn't match the requirement.
}
\description{
Checks that the input data structure `loop_points` meets
the requirements for this data structure.
}
\details{
Cross section feature classes evolve as different steps are
performed on them. The `step` parameter allows a `loop_points` data
structure to be checked throughout its lifecycle. Each step defines a
changing set of requirements for the `loop_points` data structure.
}
