% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_slope_sinuosity.R
\name{check_slope_sinuosity}
\alias{check_slope_sinuosity}
\title{Check the validity of the `slope_sinuosity` function}
\usage{
check_slope_sinuosity(slope_sinuosity_df)
}
\arguments{
\item{slope_sinuosity_df}{data frame; a data frame produced by the
`slope_sinousity` functionv.}
}
\value{
Returns TRUE if the `slope_sinuosity_df` data frame matches the
requirements. The function throws an error for a data structure not matching
the data specification. Returns errors describing how the the data structure
doesn't match the requirement.
}
\description{
Checks that the output of the `slope_sinuosity` function is
valid.
}
