% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valleyline.R
\name{check_valleyline}
\alias{check_valleyline}
\title{Check the validity of an `fluvgeo` `valleyline` data structure}
\usage{
check_valleyline(valleyline)
}
\arguments{
\item{valleyline}{sf object: a `valleyline` data structure
used by the fluvgeo package.}
}
\value{
Returns TRUE if the `valleyline` data structure matches the
requirements. The function throws an error for a data structure not matching
the data specification. Returns errors describing how the the data structure
doesn't match the requirement.
}
\description{
Checks that the input data structure `valleyline` meets
the requirements for this data structure.
}
