% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_xs_dims.R
\name{check_xs_dims}
\alias{check_xs_dims}
\title{Check the validity of an `fgm` `xs_dims` data structure}
\usage{
check_xs_dims(xs_dims)
}
\arguments{
\item{xs_dims}{data frame; a `xs_dims` data structure used
by the fgm package.}
}
\value{
Returns TRUE if the `xs_dims` data structure matches the
requirements. The function throws an error for a data structure not matching
the data specification. Returns errors describing how the the data structure
doesn't match the requirement.
}
\description{
Checks that the input data structure `xs_dims` meets
the requirements for this data structure.
}
