% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_units.R
\name{convert_distance_value}
\alias{convert_distance_value}
\title{Convert a distance from kilometers to display units}
\usage{
convert_distance_value(x, unit_system = "USCS")
}
\arguments{
\item{x}{numeric; profile distance in kilometers.}

\item{unit_system}{character; one of "USCS" or "SI".}
}
\value{
Numeric vector in display units.
}
\description{
Convert a distance from kilometers to display units
}
\keyword{internal}
