% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_units.R
\name{convert_profile_distance}
\alias{convert_profile_distance}
\title{Convert a profile distance for display}
\usage{
convert_profile_distance(x, unit_system = "USCS")
}
\arguments{
\item{x}{numeric; profile distance in kilometers.}

\item{unit_system}{character; one of "USCS" or "SI".}
}
\value{
Numeric vector converted for display.
}
\description{
Convert a profile distance for display
}
\keyword{internal}
