% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section_dimensions.R
\name{cross_section_dimensions}
\alias{cross_section_dimensions}
\title{Calculate cross section dimensions}
\usage{
cross_section_dimensions(
  xs,
  xs_points,
  bankfull_elevation,
  lead_n,
  use_smoothing,
  loess_span,
  vert_units
)
}
\arguments{
\item{xs}{SpatialLinesDataFrame; the full path to a
cross section line feature class}

\item{xs_points}{SpatialPointsDataFrame; the full path to a
cross section points feature class}

\item{bankfull_elevation}{numeric; The bankfull elevation (in feet) that is
used to calculate hydraulic geometry.}

\item{lead_n}{numeric; The number of features to lead/lag on
either side of each feature that will be used to
calculate the slope and sinuosity.}

\item{use_smoothing}{boolean; determines if smoothed elevation values
are used to calculate gradient. values are:
TRUE, FALSE (default)}

\item{loess_span}{numeric; the loess regression span parameter,
defaults to 0.05}

\item{vert_units}{character; The vertical units. One of: "m"
(meter), "ft" (foot), "us-ft" (us survey foot)}
}
\value{
A new cross section data frame with the hydraulic geometry
     dimensions added.
}
\description{
Calculates cross section hydraulic geometry dimensions for
    the input cross section feature class.
}
\examples{
# Calculate cross section dimensions
xs_dims <- cross_section_dimensions(xs = fluvgeo::sin_riffle_channel_sp,
                                    xs_points = fluvgeo::sin_riffle_channel_points_sp,
                                    bankfull_elevation = 103,
                                    lead_n = 1,
                                    use_smoothing = TRUE,
                                    loess_span = 0.5,
                                    vert_units = "ft")

}
