% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section_dimensions_L1.R
\name{cross_section_dimensions_L1}
\alias{cross_section_dimensions_L1}
\title{Calculate Cross Section Dimensions at Level 1}
\usage{
cross_section_dimensions_L1(
  xs_sf,
  lead_n,
  use_smoothing,
  loess_span,
  vert_units
)
}
\arguments{
\item{xs_sf}{sf object; the full path to a
cross section line feature class}

\item{lead_n}{numeric; The number of features to lead/lag on
either side of each feature that will be used to
calculate the slope and sinuosity.}

\item{use_smoothing}{boolean; determines if smoothed elevation values
are used to calculate gradient. values are:
TRUE, FALSE (default)}

\item{loess_span}{numeric; the loess regression span parameter,
defaults to 0.05}

\item{vert_units}{character; The vertical units. One of: "m"
(meter), "ft" (foot), "us-ft" (us survey foot)}
}
\value{
A new cross section data frame with the hydraulic geometry
     dimensions added.
}
\description{
Calculates level 1 cross section dimensions for the input
cross section feature class.
}
