% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem2rem.R
\name{dem2rem}
\alias{dem2rem}
\title{Convert DEM to REM}
\usage{
dem2rem(dem, flowline, flowline_points, buffer_distance)
}
\arguments{
\item{dem}{terra SpatRast object; A DEM.}

\item{flowline}{sf object; A flowline object.}

\item{flowline_points}{sf object; A flowline_points feature.}

\item{buffer_distance}{numeric; Distance the flowline feature class 
will be buffered to define the extent of the output 
REM. Units are defined by the coordinate system of 
the DEM.}
}
\value{
a terra SpatRaster object representing the REM
}
\description{
Converts a Digital Elevation Model (DEM) to a Relative 
              Elevation Model (REM). An REM is a DEM normalized
}
