% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend}
\alias{detrend}
\title{Detrend DEM}
\usage{
detrend(dem, flowline, flowline_points, buffer_distance)
}
\arguments{
\item{dem}{terra SpatRast object; A DEM.}

\item{flowline}{sf object; A flowline object.}

\item{flowline_points}{sf object; A flowline_points feature.}

\item{buffer_distance}{numeric; Distance the flowline feature class
will be buffered to define the extent of the output
REM. Units are defined by the coordinate system of
the DEM.}
}
\value{
a list containing two terra SpatRaster objects:
* rem   - A raster Relative Elevation Model (REM) representing elevation
          above the baseline elevation of the reach.
* trend - A raster trend surface representing the baseline elevation of
          the flowline_points for the reach.
}
\description{
Converts a Digital Elevation Model (DEM) to a Relative
              Elevation Model (REM). An REM is a DEM normalized to the
              baseline elevation of the stream reach.
}
