% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri_raster2SpatRaster.R
\name{esri_raster2SpatRaster}
\alias{esri_raster2SpatRaster}
\title{Convert an ESRI raster to a SpatRaster}
\usage{
esri_raster2SpatRaster(raster_path)
}
\arguments{
\item{raster_path}{character; A path to an ESRI raster.}
}
\value{
A `terra::SpatRaster` object.
}
\description{
Converts an ESRI raster to a terra::SpatRaster.
}
\details{
Rasters stored in ESRI Geodatabases have historically only been
accessible using the `arcgisbinding` package. No open source drivers existed
to read file geodatabase rasters into R. However, since the release of GDAL
v3.7, the OpenFileGDB driver now provides support for GDB rasters. GDAL
access in R is currently best provided using `sf`. Check
`sf::sf_extSoftVersion()` for the version of GDAL bundled with `sf`. When
`sf` migrates to using this GDAL version and the OpenFileGDB raster driver
is determined to be working, this function will migrate to using it instead
of the proprietary `arcgisbinding` approach used up to now.
}
