% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bankfull.R
\name{estimate_bankfull}
\alias{estimate_bankfull}
\title{Estimate bankfull report}
\usage{
estimate_bankfull(
  xs_points,
  streams,
  regions,
  bankfull_elevations,
  bf_estimate,
  stat,
  output_dir,
  output_format
)
}
\arguments{
\item{xs_points}{data frame; a data frame of cross section points}

\item{streams}{character vector; The stream names in the study
area.}

\item{regions}{character; The regions that a dimension will be
calculated for. See the regional_curves$region
field for a complete list.}

\item{bankfull_elevations}{numeric vector; The bankfull elevations (units:
detrended feet) that are used to calculate
hydraulic geometry.}

\item{bf_estimate}{numeric; The detrended bankfull elevation (in
feet) that is used to calculate hydraulic
geometry.}

\item{stat}{character; The statistic to graph "RMSE", "MAE"
(the default).}

\item{output_dir}{character; The output directory.}

\item{output_format}{character; The output format of the report. One
of "html_document", "word_document",
"pdf_document".}
}
\value{
A report written to the file system in the output fromat requested.
}
\description{
Produces the bankfull estimate report.
}
\examples{
# Extract attribute data from the fluvgeo::sin_xs_points SpatialPointsDataFrame
sin_xs_points_df <- fluvgeo::sin_xs_points@data

# Set variable values
streams <- c("Sinsinawa")
regions <- c("Eastern United States", "IN Central Till Plain")
bankfull_elevations <- seq(103, 104, 0.1)
bf_estimate <- 103.5
stat <- "MAE"
output_dir <- Sys.getenv("HOME")
output_format <- "html_document"

# Call the estimate_bankfull function with test data
estimate_bankfull(xs_points = sin_xs_points_df,
                  streams = streams,
                  regions = regions,
                  bankfull_elevations = bankfull_elevations,
                  bf_estimate = bf_estimate,
                  stat = stat,
                  output_dir = output_dir,
                  output_format = output_format)

}
