% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extent.R
\name{feature_extent}
\alias{feature_extent}
\title{Calculate Expanded Feature Extent}
\usage{
feature_extent(feature, extent_factor = 1)
}
\arguments{
\item{feature}{sf; A feature class.}

\item{extent_factor}{numeric; A numeric value used to expand the extent
of the supplied feature class.}
}
\value{
a raster::extent object
}
\description{
Creates a new expanded map extent based on a supplied
map extent multiplying factor.
}
