% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{fgm}
\alias{fgm}
\alias{fgm-package}
\title{A Package for Fluvial Geomorphic Analysis}
\description{
This package contains a wide range of functions for performing
    fluvial geomorphic analysis. This package is designed for use with the 
    FluvialGeomorph Python package. The FluvialGeomorph package contains an 
    ArcGIS toolbox that can be used to extract spatial data from high-
    resolution terrain data. FluvialGeomorph uses functions in this package 
    to perform a wide range of fluvial geomorphic analysis (e.g., extract 
    stream channel dimensions, help choose bankfull elevation for ungaged 
    streams).
}
\details{
Place some other package details here. And then some more.
}
\section{Installation}{

The functions \code{\link{arc2sp}} and \code{\link{sp2arc}} require the
prior installation of the \code{arcgisbinding} package AND a licensed
installation of the commercial GIS software ESRI \code{ArcGIS Desktop}
or \code{ArcGIS Pro}. The \code{arcgisbinding} package can be installed
from within \code{ArcGIS Pro} or by following the instructions at
\url{https://r-arcgis.github.io/assets/arcgisbinding-vignette.html} for
\code{ArcGIS Desktop} users. No other functions within this package
require ESRI \code{ArcGIS Desktop} or \code{ArcGIS Pro}.
}

\references{
\describe{
  \item{ESRI ArcGIS Desktop, ArcGIS Pro}{\url{https://pro.arcgis.com/}}
  \item{\code{arcgisbinding}}{
  \url{https://r-arcgis.github.io/assets/arcgisbinding-vignette.html}}
}
}
\author{
\strong{Maintainer}: Michael Dougherty \email{Michael.P.Dougherty@usace.army.mil}

Authors:
\itemize{
  \item Christopher Haring \email{Christopher.P.Haring@usace.army.mil}
  \item Charles Theiling \email{Charles.H.Theiling@usace.army.mil}
}

Other contributors:
\itemize{
  \item Davi Michl [contributor]
  \item Thomas Darby [contributor]
}

}
\keyword{internal}
