% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_xs_profiles_L1.R
\name{fig_xs_profiles_L1}
\alias{fig_xs_profiles_L1}
\title{Cross Section Plot Figure}
\usage{
fig_xs_profiles_L1(
  cross_section,
  xs_number,
  dem,
  channel = NULL,
  floodplain = NULL,
  extent_factor = 1,
  xs_pts_sf_list
)
}
\arguments{
\item{cross_section}{sf; A cross section lines feature class.}

\item{xs_number}{integer; The cross section `Seq` number of the
requested cross section.}

\item{dem}{terra SpatRaster; A dem raster.}

\item{channel}{sf; A channel polygon feature class (optional).}

\item{floodplain}{sf; A floodplain polygon feature class (optional).}

\item{extent_factor}{numeric; A numeric value used to expand the map
extent around each cross section.'}

\item{xs_pts_sf_list}{list; a list of `sf` objects of cross
section points, one for each survey time period
to be graphed. Survey list items must be tagged
with the survey label to be used in the graph
legend.}
}
\value{
A patchwork figure.
}
\description{
Constructs a figure of plots describing a single cross section.
}
