% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodplain_volume.R
\name{floodplain_volume}
\alias{floodplain_volume}
\title{Floodplain Volume}
\usage{
floodplain_volume(dem, watersurface)
}
\arguments{
\item{dem}{terra SpatRast object; A DEM. Elevation units = feet.}

\item{watersurface}{terra SpatRaster object; The water surface. Created
using the detrend$trend surface and adding water depth
in "bathtub" inundation modeling.}
}
\value{
A numeric value representing the volume between the teo surfaces
in cubic meters.
}
\description{
Calculate the volume of the floodplain  between a DEM and a
             water surface.
}
