% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluvialGeomorphicMetric.R
\docType{class}
\name{FluvialGeomorphicMetric-class}
\alias{FluvialGeomorphicMetric-class}
\title{An S4 class to represent a fluvial geomorphic metric}
\description{
An S4 class to represent a fluvial geomorphic metric
}
\section{Slots}{

\describe{
\item{\code{metric}}{name of the metric}

\item{\code{definition}}{mathmatical definition of the metric}

\item{\code{threshold_breaks}}{numeric breaks of thresholds}

\item{\code{threshold_labels}}{text description of the catagories defined by
threshold_breaks}

\item{\code{source}}{citation for the thresholds}
}}

