% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_slash.R
\name{forward_slash}
\alias{forward_slash}
\title{Convert Windows path to forward slashes}
\usage{
forward_slash(path)
}
\arguments{
\item{path}{character; A windows file path containing escaped
backslashs (i.e., \\)}
}
\description{
Convert a Windows file path from double backslashes to single
forward slash file seperators suitable for use in R.
}
