% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdb_raster2SpatRast.R
\name{gdb_raster2SpatRast}
\alias{gdb_raster2SpatRast}
\title{Convert an ESRI Geodatabase raster to a SpatRaster}
\usage{
gdb_raster2SpatRast(raster_path)
}
\arguments{
\item{raster_path}{character; A path to an ESRI raster.}
}
\value{
A `terra::SpatRaster` object.
}
\description{
Converts an ESRI Geodatabase raster to a terra::SpatRaster.
}
