% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arc_wkt.R
\name{get_arc_wkt}
\alias{get_arc_wkt}
\title{Get ArcGIS feature class CRS WKT}
\usage{
get_arc_wkt(fc_path, WKID = FALSE)
}
\arguments{
\item{fc_path}{character; Path to the ArcGIS geodatabase feature class.}

\item{WKID}{logical; TRUE returns the CRS WKT string (default),
FALSE returns the CRS WKID integer.}
}
\value{
CRS WKT string.
}
\description{
Get ArcGIS feature class coordinate reference system (CRS) WKT
  string.
}
