% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{get_dem}
\alias{get_dem}
\title{Get DEM}
\usage{
get_dem(xs)
}
\arguments{
\item{xs}{sf object}
}
\value{
terra::SpatRaster
}
\description{
Get a DEM from the ESRI WorldElevation Terrain Image Service
that covers the extent of the input features.
}
