% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leaflet.R
\name{get_leaflet}
\alias{get_leaflet}
\title{Get Leaflet Map}
\usage{
get_leaflet(search = FALSE)
}
\arguments{
\item{search}{logical; Display a search bar?}
}
\value{
a leaflet map object
}
\description{
Creates a leaflet map.
}
