% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terrain_leaflet.R
\name{get_terrain_leaflet}
\alias{get_terrain_leaflet}
\title{Get a Terrain Leaflet}
\usage{
get_terrain_leaflet(xs, dem)
}
\arguments{
\item{xs}{sf object, The cross sections}

\item{dem}{SpatRaster object, The digital elevation model.}
}
\value{
a leaflet object
}
\description{
Get a terrain map as a leaflet object.
}
