% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_watershed.R
\name{get_watershed}
\alias{get_watershed}
\title{Get Watershed}
\usage{
get_watershed(pour_point)
}
\arguments{
\item{pour_point}{sf object; An sf pour point object.}
}
\value{
An sf polygon object representing the upstream drainage area of
         the pour point.
}
\description{
Get watershed from ESRI geoprocessing service.
}
