% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_graph.R
\name{gof_graph}
\alias{gof_graph}
\title{Plot goodness-of-fit graph}
\usage{
gof_graph(gof_stats, stream, bankfull_elevation, stat = "MAE")
}
\arguments{
\item{gof_stats}{data frame; A gof_stats data frame. See the
\code{\link{build_gof_stats}} function for
details on how to create.}

\item{stream}{character; The name of the stream.}

\item{bankfull_elevation}{numeric; The detrended bankfull elevation (in
feet) that is used to calculate hydraulic
geometry.}

\item{stat}{character; The statistic to graph "RMSE", "MAE"
(the default).}
}
\value{
A ggplot object.
}
\description{
Produces a goodness-of-fit statistic graph for the specified
reach for analysis regions specified by the \code{\link{build_gof_stats}}
function.
}
\details{
Add a discussion of how to interpret the graph.
}
\examples{
# Extract attribute data from the fgm::sin_xs_points SpatialPointsDataFrame
sin_xs_points_df <- fgm::sin_riffle_channel_points_sp@data

# Set variable values
streams <- c("Sinsinawa")
regions <- c("Eastern United States", "IN Central Till Plain")
bankfull_elevations = seq(103, 104, 0.1)

# Call the xs_dimensions function
sin <- xs_dimensions(xs_points = sin_xs_points_df,
                     streams = streams,
                     regions = regions,
                     bankfull_elevations = bankfull_elevations)

# Call the build_gof_stats function
sin_gof <- build_gof_stats(xs_dims = sin,
                           streams = streams,
                           regions = regions,
                           bankfull_elevations = bankfull_elevations)

# Call the gof_graph function
sin_gof_graph <- gof_graph(gof_stats = sin_gof,
                           stream = streams,
                           bankfull_elevation = 103.5,
                           stat = "MAE")

# Print the graph
sin_gof_graph

}
\seealso{
The \code{\link{build_gof_stats}} function for details on how to
create a \code{gof_stats} data frame.
}
