% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gte.R
\name{gte}
\alias{gte}
\title{Assign Vector Values Greater Than or Equal to Value}
\usage{
gte(x, value)
}
\arguments{
\item{x}{numeric vector; The vector of values to be adjusted.}

\item{value}{numeric;}
}
\value{
Returns a numeric vector greater than or equal to the `value`
parameter. Sets the values less than `value` to the value of `value`.
}
\description{
Returns a numeric vector greater than or equal to the `value`
parameter. Sets the values less than `value` to the value of `value`.
}
\details{
Used to limit the range of values from a FluvialGeomorph metric to
ensure that only conceptually relevant values are returned. This is simply a
convenience function that uses a more computationally efficient method than
the available `ifelse` functions.
}
