% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justify_gtable.R
\name{justify_gtable}
\alias{justify_gtable}
\title{Justify a tableGrob}
\usage{
justify_gtable(x, hjust = "center", vjust = "center", draw = FALSE)
}
\arguments{
\item{x}{gtable; A gtable returned from `gridExtra::tableGrob`.}

\item{hjust}{character; Horizontal justification. One of "left", "right",
or "center" (default).}

\item{vjust}{character; Vertical justification. One of "left", "right",
or "center" (default).}

\item{draw}{boolean; Should the table be drawn?}
}
\value{
A `gtable` object
}
\description{
Correctly justifies a `gtable` object returned from the
`gridExtra::tableGrob` function.
}
\details{
This function was provided in https://stackoverflow.com/a/32111143/7454101
by the `gridExtra` author baptiste.
}
