% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_1_report.R
\name{level_1_report}
\alias{level_1_report}
\title{Level 1 Report}
\usage{
level_1_report(
  stream,
  flowline_fc,
  cross_section_fc,
  flowline_points_1,
  flowline_points_2,
  flowline_points_3,
  flowline_points_4,
  xs_points_1,
  xs_points_2,
  xs_points_3,
  xs_points_4,
  survey_name_1,
  survey_name_2,
  survey_name_3,
  survey_name_4,
  features_fc,
  dem,
  show_xs_map = FALSE,
  profile_units,
  aerial = TRUE,
  elevation = FALSE,
  xs_label_freq = 5,
  exaggeration = 10,
  extent_factor = 1.2,
  output_dir,
  output_format
)
}
\arguments{
\item{stream}{character; The stream name. The stream name must
match a stream name in `ReachName` field in the
other parameters.}

\item{flowline_fc}{character; The path to a `flowline` feature class.}

\item{cross_section_fc}{character; The path to a cross section feature
class.}

\item{flowline_points_1}{character; The path to a `flowline_points` feature
class for the first time period.}

\item{flowline_points_2}{character; The path to a `flowline_points` feature
class for the second time period.}

\item{flowline_points_3}{character; The path to a `flowline_points` feature
class for the third time period.}

\item{flowline_points_4}{character; The path to a `flowline_points` feature
class for the fourth time period.}

\item{xs_points_1}{character; The path to a `xs_points` feature
class for the first time period.}

\item{xs_points_2}{character; The path to a `xs_points` feature
class for the second time period.}

\item{xs_points_3}{character; The path to a `xs_points` feature
class for the third time period.}

\item{xs_points_4}{character; The path to a `xs_points` feature
class for the fourth time period.}

\item{survey_name_1}{character: The name or date of the first survey.}

\item{survey_name_2}{character: The name or date of the second survey.}

\item{survey_name_3}{character: The name or date of the third survey.}

\item{survey_name_4}{character: The name or date of the fourth survey.}

\item{features_fc}{character; The path to a `features` feature class.}

\item{dem}{character; The path to the DEM raster.}

\item{show_xs_map}{logical; Add the cross section maps to the report?}

\item{profile_units}{character; The units to be used for the x-axis of
the longitudinal profile graphs. One of "feet",
"miles", "meters", "kilometers".}

\item{aerial}{logical; Display an overview map with an aerial
photo background?}

\item{elevation}{logical; Display an overview map with an elevation
background?}

\item{xs_label_freq}{numeric; An integer indicating the frequency of
cross section labels.}

\item{exaggeration}{numeric; The degree of terrain exaggeration.}

\item{extent_factor}{numeric; The amount the extent is expanded around
the cross section feature class. Values greater
than one zoom out, values less than one zoom in.}

\item{output_dir}{character; The path to the folder in which to
write the report.}

\item{output_format}{character; The file format of the report. One of
"html_document", "word_document", "pdf_document".}
}
\value{
Produces a FluvialGeomorph Level 1 Report in the `output_dir` in the
requested file format.
}
\description{
Creates a FluvialGeomorph Level 1 report.
}
