% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_2_report.R
\name{level_2_report}
\alias{level_2_report}
\title{Level 2 Report}
\usage{
level_2_report(
  stream,
  flowline_fc,
  xs_fc,
  xs_points_fc,
  xs_dims_fc,
  dem,
  banklines_fc,
  features_fc,
  bf_estimate,
  regions,
  extent_factor,
  label_xs,
  profile_units,
  output_dir,
  output_format
)
}
\arguments{
\item{stream}{character; The stream name. The stream name must
match a stream name in `ReachName` field in the
other parameters.}

\item{flowline_fc}{character; The path to a `flowline` feature class.}

\item{xs_fc}{character; The path to the cross section feature
class.}

\item{xs_points_fc}{character; The path to a `xs_points` feature
class for the first time period.}

\item{xs_dims_fc}{character; The path to the "xs_dims" feature class.}

\item{dem}{character; The path to the DEM raster.}

\item{banklines_fc}{character: the path to the banklines feature class.}

\item{features_fc}{character; The path to a `features` feature class.}

\item{bf_estimate}{numeric; Detrended bankfull estimate (units:
detrended feet).}

\item{regions}{character vector; Regions to calculate hydraulic
dimensions for. See the `RegionalCurve` package for
a list of regions.}

\item{extent_factor}{numeric; The extent factor used to control the
extent of cross section site maps.}

\item{label_xs}{logical; Label cross sections?}

\item{profile_units}{character; the units of the longitudinal profile.
One of "kilometers", "meters", "miles", or "feet"}

\item{output_dir}{character; The path to the folder in which to
write the report.}

\item{output_format}{character; The file format of the report. One of
"html_document", "word_document", "pdf_document".}
}
\value{
Produces a FluvialGeomorph Level 2 Report in the `output_dir` in the
requested file format.
}
\description{
Creates a level 2 report.
}
