% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reach_metric.R
\name{map_reach_metric}
\alias{map_reach_metric}
\title{Reach Metric Map}
\usage{
map_reach_metric(metric, flowline, xs_dimensions)
}
\arguments{
\item{metric}{MetricThreshold object; the fluvial geomorphic
metric to be mapped}

\item{flowline}{SpatialLinesDataFrame; a flowline feature class}

\item{xs_dimensions}{SpatialLinesDataFrame; a cross section dimensions
feature class}
}
\value{
a tmap object
}
\description{
Produces a reach fluvial geomorphic metric thematic map
  displaying the metric values at each cross section.
}
