% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reach_overview.R
\name{map_reach_overview}
\alias{map_reach_overview}
\title{Reach Overview Map}
\usage{
map_reach_overview(
  flowline_sf,
  cross_section_sf,
  background = "aerials",
  exaggeration = 20
)
}
\arguments{
\item{flowline_sf}{sf object; A flowline feature class}

\item{cross_section_sf}{sf object; A cross section feature class}

\item{background}{character; The type of map background. One of
"aerial" or "elevation"}

\item{exaggeration}{numeric; The degree of terrain exaggeration.}
}
\value{
a tmap object
}
\description{
Produces a reach overview map displaying cross section
locations over an aerial image.
}
\examples{

}
