% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_xs.R
\name{map_xs}
\alias{map_xs}
\title{Cross section location map}
\usage{
map_xs(cross_section, xs_number, dem, banklines = NULL, extent_factor = 1)
}
\arguments{
\item{cross_section}{sf; A cross section lines
feature class.}

\item{xs_number}{integer; The cross section identifier of the
requested cross section.}

\item{dem}{terra SpatRaster; A dem raster.}

\item{banklines}{sf; A banklines feature
class (optional).}

\item{extent_factor}{numeric; A numeric value used to expand the map
extent around each cross section.}
}
\value{
a tmap object
}
\description{
Produces a cross section location map for the specified cross
section.
}
