% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meander_length.R
\name{meander_length}
\alias{meander_length}
\title{Calculate the meander length for each loop}
\usage{
meander_length(bankline_points)
}
\arguments{
\item{bankline_points}{SpatialPointsDataFrame; an fgm bankline_points data
structure}
}
\value{
Returns a data frame of loops with the calculated meander length.
}
\description{
Calculates the meander length (wavelength) for each loop in the
input `bankline_points` data frame.
}
