% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_colors.R
\name{metric_colors}
\alias{metric_colors}
\title{Metric colors}
\usage{
metric_colors(metric)
}
\arguments{
\item{metric}{MetricThreshold object; the fluvial geomorphic
metric to be mapped}
}
\value{
a vector of colors
}
\description{
Returns a vector of colors based on the number of threshold
  breaks in the input `FluvialGeomorphicMetric`` object.
}
