% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planform_dimensions.R
\name{planform_dimensions}
\alias{planform_dimensions}
\title{Calculate planform dimensions for each cross section}
\usage{
planform_dimensions(xs_dimensions, bankline_points)
}
\arguments{
\item{xs_dimensions}{sf data frame; A `cross_section_dimensions`
fluvgeo data structure.}

\item{bankline_points}{sf data frame; A `bankline_points` fluvgeo data
data structure.}
}
\value{
Returns the input xs_dims data frame of cross sections with the
  calculated metric ratio variables added.
}
\description{
Calculates planform dimensions for each cross section in the
  input xs_dims data frame.
}
\examples{
# Calculate cross section planform dimensions
xs_dims_plan <- planform_dimensions(fluvgeo::sin_riffle_floodplain_dims_L2_sf,
                                   fluvgeo::sin_bankline_points_sf)
}
