% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_watershed_area.R
\name{pt_watershed_area}
\alias{pt_watershed_area}
\title{Point Watershed Area}
\usage{
pt_watershed_area(point_sf)
}
\arguments{
\item{point_sf}{an sf object; The point to have a watershed determined.}
}
\value{
a numeric value of the watershed area in units square miles
}
\description{
Calculates the upstream watershed area for the specified point.
}
