% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_metric_map.R
\name{reach_metric_map}
\alias{reach_metric_map}
\title{Reach Metric Map}
\usage{
reach_metric_map(metric, flowline, xs_dimensions)
}
\arguments{
\item{metric}{MetricThreshold object; the fluvial geomorphic
metric to be mapped}

\item{flowline}{SpatialLinesDataFrame; a flowline feature class}

\item{xs_dimensions}{SpatialLinesDataFrame; a cross section dimensions
feature class}
}
\value{
a tmap object
}
\description{
Produces a reach fluvial geomorphic metric thematic map
  displaying the metric values at each cross section.
}
