% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_overview_map.R
\name{reach_overview_map}
\alias{reach_overview_map}
\title{Reach Overview Map}
\usage{
reach_overview_map(flowline, xs_dimensions)
}
\arguments{
\item{flowline}{SpatialLinesDataFrame; a flowline feature class}

\item{xs_dimensions}{SpatialLinesDataFrame; a cross section dimensions
feature class}
}
\value{
a tmap object
}
\description{
Produces a reach overview map displaying cross section
locations over an aerial image.
}
