% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2arc.R
\name{sf2arc}
\alias{sf2arc}
\title{Converts an \code{sf} object to an ArcGIS feature class}
\usage{
sf2arc(sf_object, fc_path)
}
\arguments{
\item{sf_object}{\code{sf} object}

\item{fc_path}{character; Path to the ArcGIS output feature class. The
feature class must reside in a file geodatabase feature
dataset.}
}
\value{
Writes the \code{sf} object to an ArcGIS feature class specified
by fc_path.
}
\description{
Converts an \code{sf} object to an ArcGIS feature class in
    the specified geodatabase feature class.
}
\details{
This function requires the prior installation of the
\code{arcgisbinding} package AND a licensed installation of ESRI
\code{ArcGIS Pro}. The \code{arcgisbinding} package can be installed
from within \code{ArcGIS Pro}.
}
\references{
\describe{
  \item{ArcGIS Pro}{\url{https://pro.arcgis.com/}}
  \item{\code{arcgisbinding}}{
  \url{https://r-arcgis.github.io/assets/arcgisbinding-vignette.html}}
}
}
\seealso{
This function for loading data from an ESRI spatial dataset.
}
