% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2arc_table.R
\name{sf2arc_table}
\alias{sf2arc_table}
\title{Save `sf` object to an ArcGIS table.}
\usage{
sf2arc_table(sf_object, table_path)
}
\arguments{
\item{sf_object}{`sf` object; The object whose data will be saved
as a table.}

\item{table_path}{character; Path to the ArcGIS output geodatabase table.}
}
\value{
Writes the `sf_object` as a geodatabase table specified by
`table_path`
}
\description{
Save `sf` object to an ArcGIS table.
}
