% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2csv.R
\name{sf2csv}
\alias{sf2csv}
\title{Save `sf` object to a .csv file.}
\usage{
sf2csv(sf_object, csv_path)
}
\arguments{
\item{sf_object}{`sf` object; The object whose data frame will be saved
as a csv file.}

\item{csv_path}{character; Path to the output csv file.}
}
\value{
Writes the `sf_object` as a csv file specified by
`csv_path`
}
\description{
Save `sf` object to a .csv file.
}
