% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_line_end_point.R
\name{sf_line_end_point}
\alias{sf_line_end_point}
\title{Extract line end points}
\usage{
sf_line_end_point(line, end)
}
\arguments{
\item{line}{sf line object, The input line feature.
be calculated.}

\item{end}{character, Which end of the line? One of "start" or "end".}
}
\value{
sf point object with x and y coordinate value attributes
}
\description{
Determines the requested line end and returns a new sf point
object with x and y coordinate value attributes added. Also adds ReachName
XS Seq attributes suitable for joining.
}
