% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_line_reverse.R
\name{sf_line_reverse}
\alias{sf_line_reverse}
\title{Reverse line direction}
\usage{
sf_line_reverse(line_sf)
}
\arguments{
\item{line_sf}{sf object of type MULTILINESTRING, LINESTRING}
}
\value{
sf object
}
\description{
Reverses the directions of an sf object's lines.
}
