% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shear_stress.R
\name{shear_stress}
\alias{shear_stress}
\title{Calculate shear stress for each cross section}
\usage{
shear_stress(xs_dims)
}
\arguments{
\item{xs_dims}{data frame; a data frame of cross section dimensions}
}
\value{
Returns a data frame of cross sections with the calculated shear
stress.
}
\description{
Calculates the shear stress for each cross section in the
input xs_dims data frame.
}
