% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_bankline_points_sf.R
\docType{data}
\name{sin_bankline_points_sf}
\alias{sin_bankline_points_sf}
\title{Sinsinawa Creek, IL stream bankline points}
\format{
A \code{sf} with 1413 observations and 9
    variables.
\describe{
    \item{OBJECTID}{numeric; The ArcGIS feature class assigned unique
                    identifier for each point in the dataset.}
    \item{ReachName}{character; The name of the stream. }
    \item{bank_id}{numeric; The bank unique identifier. Right descending = 1,
                   left descending = 2. }
}
}
\source{
This dataset was produced using the \code{FluvialGeomorph} ArcGIS
toolbox.
}
\usage{
sin_bankline_points_sf
}
\description{
A `sf` object containing stream
bankline points for Sinsinawa Creek, IL. These points represent the bankfull
bankline extracted from a LiDAR terrain surface.
}
\keyword{datasets}
