% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_banklines_sf.R
\docType{data}
\name{sin_banklines_sf}
\alias{sin_banklines_sf}
\title{Sinsinawa Creek, IL stream banklines}
\format{
A \code{sf} with 2 observations and 7
    variables.
\describe{
    \item{OBJECTID}{numeric; The ArcGIS feature class assigned unique
                    identifier for each point in the dataset.}
    \item{ReachName}{character; The name of the stream. }
}
}
\source{
This dataset was produced using the \code{FluvialGeomorph} ArcGIS
toolbox.
}
\usage{
sin_banklines_sf
}
\description{
A \code{sf} object containing stream
banklines for Sinsinawa Creek, IL. These lines represent the bankfull
banklines extracted from a LiDAR terrain surface.
}
\keyword{datasets}
