% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_features_sp.R
\docType{data}
\name{sin_features_sp}
\alias{sin_features_sp}
\title{Sinsinawa Creek, IL stream features}
\format{
A \code{SpatialPointsDataFrame} with 5 observations and 4
    variables.
\describe{
    \item{OBJECTID}{numeric; The ArcGIS feature class assigned unique
                    identifier for each point in the dataset.}
    \item{Name}{character; The name of the stream feature. }
    \item{km_to_mouth}{numeric; The longitudinal position of the feature.}
}
}
\source{
This dataset was produced using the \code{FluvialGeomorph} ArcGIS
toolbox.
}
\usage{
sin_features_sp
}
\description{
A \code{sp::SpatialPointsDataFrame} object containing stream
feature points used to identify and label key stream features and their
longitudinal position. Instances of features typically include relevant
tributaries, localities within the floodplain affecting flow.
}
\keyword{datasets}
