% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_flowline_points_sf.R
\docType{data}
\name{sin_flowline_points_sf}
\alias{sin_flowline_points_sf}
\title{Sinsinawa Creek, IL stream flowline points}
\format{
A \code{sf} with 1413 observations and 9
    variables.
\describe{
    \item{OBJECTID}{numeric; The ArcGIS feature class assigned unique
                    identifier for each point in the dataset.}
    \item{ReachName}{character; The name of the stream. }
    \item{POINT_X}{numeric; The longitude or easting of the point.
                   Coordinate system units can be determined by checking the
                   \code{st_crs(sin_xs_points)}.}
    \item{POINT_Y}{numeric; The latitude or northing of the point.
                   Coordinate system units can be determined by checking the
                   \code{st_crs(sin_xs_points)}.}
    \item{POINT_M}{numeric; The route distance of this cross section from
                   the upstream end of the reach. Coordinate system units
                   can be determined by checking the
                   \code{st_crs(sin_xs_points)}.}
    \item{Z}{numeric; The elevation of the point. By convention, the
                 vertical units used in this project are NADV88 feet.}
}
}
\source{
This dataset was produced using the \code{FluvialGeomorph} ArcGIS
toolbox.
}
\usage{
sin_flowline_points_sf
}
\description{
A \code{sf} object containing stream
flowline points for Sinsinawa Creek, IL. These points represent elevation
values extracted from a LiDAR terrain surface along the flowline of the
stream.
}
\keyword{datasets}
