% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_left.R
\name{start_left}
\alias{start_left}
\title{Does point C fall on the left side of the AB line?}
\usage{
start_left(a_x, a_y, b_x, b_y, c_x, c_y)
}
\arguments{
\item{a_x}{numeric, Point A x coordinate}

\item{a_y}{numeric, Point A y coordinate}

\item{b_x}{numeric, Point B x coordinate}

\item{b_y}{numeric, Point B y coordinate}

\item{c_x}{numeric, Point C x coordinate}

\item{c_y}{numeric, Point C y coordinate}
}
\value{
logical
}
\description{
Returns TRUE if the point C fall on the left side of the AB
line. Orientation of the AB line determines the left and right side (i.e,
starting at point A looking toward point B). Used to answer the question:
Is point C on the left of the AB line?
}
\details{
Uses the cross product method. Taken from:
[SO](https://stackoverflow.com/questions/1560492/how-to-tell-whether-a-point-is-to-the-right-or-left-side-of-a-line)
}
