% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_power.R
\name{stream_power}
\alias{stream_power}
\title{Calculate stream power for each cross section}
\usage{
stream_power(xs_dims, discharge_method = c("model_measure",
  "regional_curve", "width_relationship"), discharge_value = NULL,
  region = NULL, drainage_area = NULL, width_method = NULL)
}
\arguments{
\item{xs_dims}{data frame; A data frame of cross section
dimensions.}

\item{discharge_method}{character; The method for calculating discharge (Q).
Must be one of: "model_measure", "regional_curve",
"width_relationship".}

\item{discharge_value}{numeric; The discharge value (single value or vector)
to use for the stream power calculation. Required
if discharge_method = "model_measure".}

\item{region}{character; The regional curve name used to calculate
discharge. Required if discharge_method =
"regional_curve". This parameter is passed to the
RegionalCurve::RHG function. See the RegionalCurve
package for a list of regions with discharge
relationships.}

\item{drainage_area}{numeric; The drainage area (single value or vector)
used by the RegionalCurve::RHG function to calculate
discharge. Required if discharge_method =
"regional_curve".}

\item{width_method}{character; The name of the width relationship used
to calculate discharge (Q) from width. Required if
discharge_method = "width_relationship". Must be
one of: <list goes here when implemented>}
}
\value{
Returns the input xs_dims data frame of cross sections with the
  calculated stream power variables added.
}
\description{
Calculates stream power for each cross section in the input
  xs_dims data frame.
}
\details{
\strong{Stream Power Equations}
  Stream power is the rate of energy dissipation against the bed and banks
  of a river or stream per unit downstream length. It is given by the
  equation:

  \deqn{\Omega = \rho g Q S}

  where \eqn{\Omega} is the stream power, \eqn{\rho} is the density of water
  (assumes 1000 kg/m3), \eqn{g} is acceleration due to gravity (9.8 m/s2),
  \eqn{Q} is discharge (m3/s), and \eqn{S} is the channel slope.

  Unit stream power is stream power per unit channel width, and is given by
  the equation:

  \deqn{\omega = (\rho g Q S) / b}

  where \eqn{\omega} is the unit stream power, and \eqn{b} is the width of
  the channel (m).

  \url{https://en.wikipedia.org/wiki/Stream_power}

  \strong{Lane's Balance Equation}
  Stream power in the Lane's Balance equation is simply represented by
  discharge \eqn{Q} and slope \eqn{S}:

  \deqn{\Omega = Q S}

  \strong{Determining Discharge}
  The \code{discharge_method} parameter specifies the source used for the
  discharge factor in the stream power equations.

  \itemize{
    \item \code{model_measure} Use this option to supply an estimate of
    discharge from a hydraulic model or field measurement.
    \item \code{regional_curve} Use this option to use an estimate of
    discharge from a regional curve.
    \item \code{width_relationship} Use this option to use an estimate of
    discharge from a width relationship
  }
}
\examples{
# Calculate cross section dimensions
xs_dims <- cross_section_dimensions(xs = fgm::sin_riffle_channel_sp,
                             xs_points = fgm::sin_riffle_channel_points_sp,
                             bankfull_elevation = 103,
                             lead_n = 1,
                             use_smoothing = TRUE,
                             loess_span = 0.5)

# Calculate stream power
xs_dims_sp <- stream_power(xs_dims,
                           discharge_method = "regional_curve",
                           region = "Illinois River",
                           drainage_area = 41)

}
