% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sx2arc_table.R
\name{sx2arc_table}
\alias{sx2arc_table}
\title{Save `sf` or `sp` object to an ArcGIS table.}
\usage{
sx2arc_table(sx_object, table_path)
}
\arguments{
\item{sx_object}{`sp` or `sf` object; The object whose data will be saved
as a table.}

\item{table_path}{character; Path to the ArcGIS output geodatabase table.}
}
\value{
Writes the `sx_object` as a geodatabase table specified by
`table_path`
}
\description{
Save `sf` or `sp` object to an ArcGIS table.
}
