% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_xs_dimensions.R
\name{table_xs_dimensions}
\alias{table_xs_dimensions}
\title{Create a Cross Section Dimensions Table}
\usage{
table_xs_dimensions(xs_pts_sf, xs_number, bf_estimate, regions)
}
\arguments{
\item{xs_pts_sf}{sf; A cross section lines feature class.}

\item{xs_number}{integer; The cross section `Seq` number of the
requested cross section.}

\item{bf_estimate}{numeric; Detrended bankfull estimate (units:
detrended feet).}

\item{regions}{character vector; Regions to calculate hydraulic
dimensions for. See the `RegionalCurve` package for
a list of regions.}
}
\value{
a `gtable` object
}
\description{
Creates a cross section dimensions table for the channel portion of the
specified cross section and regions using the `RegionalCurve` R package.
}
