% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_stream_power.R
\name{unit_stream_power}
\alias{unit_stream_power}
\title{Calculates unit stream power}
\usage{
unit_stream_power(reach_xs_dims)
}
\arguments{
\item{reach_xs_dims}{data frame; a data frame of cross section
dimensions.}
}
\value{
Returns a data frame of loops with the calculated meander width.
}
\description{
Calculates the unit stream power for the input cross sections.
}
\examples{
# Extract data from the fgm::sin_xs_dimensions SpatialPointsDataFrame
sin_xs_dims_df <- fgm::sin_xs_dimensions@data



}
