% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_gof_stats.R
\name{update_gof_stats}
\alias{update_gof_stats}
\title{Update the \code{gof_stats} data frame}
\usage{
update_gof_stats(reach_name, xs_dims, gof_stats, regions,
  bankfull_elevations)
}
\arguments{
\item{reach_name}{character; The reach name of the new xs_points data.}

\item{xs_dims}{data frame; The data frame that holds cross section
dimensions.}

\item{gof_stats}{data frame; The data frame that holds the goodnesss-
of-fit statistics.}

\item{regions}{character vector; A vector of region names for which
GOF stats will be calculated.}

\item{bankfull_elevations}{numeric vector; A numeric vector of detrended
bankfull elevations (units feet) for which GOF
stats will be calculated.}
}
\value{
An updated \code{gof_stats} data frame for the specifed reach.
}
\description{
Updates the gof_stats data frame with goodness-of-fit scores
    for cross section dimensions for the specified reach.
}
