% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_reach.R
\name{update_reach}
\alias{update_reach}
\title{Update dimensions and stats for the input reach}
\usage{
update_reach(xs_points_path, xs_points, xs_dims, gof_stats, regions,
  bankfull_elevations)
}
\arguments{
\item{xs_points_path}{character; The path to the file geodatabase feature
class "cross_section_points" for the specified site.}

\item{xs_points}{sp object; The sp object containing the xs_points for
the study area.}

\item{xs_dims}{data frame; The data frame that holds cross section
dimensions.}

\item{gof_stats}{data frame; The data frame that holds the gof stats.}

\item{regions}{character vector; A vector of region names for which
GOF stats will be calculated.}

\item{bankfull_elevations}{numeric vector; A numeric vector of bankfull
elevations for which GOF stats will be
calculated.}
}
\value{
A list containing three data frames updated for the specified reach:
    1. xs_dims, 2. xs_points, 3. gof_stats
}
\description{
Update cross section dimensions and GOF stats for the input
    reach.
}
