% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_xs_dims.R
\name{update_xs_dims}
\alias{update_xs_dims}
\title{Update the \code{xs_dims} data frame}
\usage{
update_xs_dims(reach_name, xs_points_reach, xs_dims, regions,
  bankfull_elevations)
}
\arguments{
\item{reach_name}{character; The reach name of the new xs_points data.}

\item{xs_points_reach}{sp object; The sp object containing the xs_points.}

\item{xs_dims}{data frame; The data frame that holds cross section
dimensions.}

\item{regions}{character vector; A vector of region names for which
GOF stats will be calculated.}

\item{bankfull_elevations}{numeric vector; A numeric vector of detrended
bankfull elevations for which GOF stats will be
calculated.}
}
\value{
An updated data frame of xs_dims for the specified reach.
}
\description{
Updates the xs_dims data frame with cross section dimensions
    for the specified reach.
}
