% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_xs_points.R
\name{update_xs_points}
\alias{update_xs_points}
\title{Update the \code{xs_points} data frame}
\usage{
update_xs_points(reach_name, xs_points_reach, xs_points)
}
\arguments{
\item{reach_name}{character; The reach name of the new xs_points
data.}

\item{xs_points_reach}{\code{sp} object; The \code{sp} object containing
the \code{xs_points} for the specified reach.}

\item{xs_points}{\code{sp} object; The \code{sp} object containing
the \code{xs_points} for the study area.}
}
\value{
An updated SpatialPointsDataFrame object containing the terrain data
    for the specified reach.
}
\description{
Updates the \code{xs_points} SpatialPointsDataFrame object
    with terrain data for the specified reach.
}
