% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_surface_poly.R
\name{water_surface_poly}
\alias{water_surface_poly}
\title{Water Surface Polygon}
\usage{
water_surface_poly(rem, water_surface_elevation, flowline)
}
\arguments{
\item{rem}{terra SpatRaster object; An REM raster.}

\item{water_surface_elevation}{numeric; The REM elevation of the water
surface.}

\item{flowline}{sf object; A flowline object.}
}
\value{
An sf polygon object representing the the water surface extent at
         the specified REM elevation.
}
\description{
Derives a water surface extent polygon from an REM at the
             specified REM elevation.
}
