% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs2pts.R
\name{xs2pts}
\alias{xs2pts}
\title{XS lines to points}
\usage{
xs2pts(xs_sp)
}
\arguments{
\item{xs_sp}{SpatialLinesDataFrame; a cross section data structure
created by the fluvgeo package.}
}
\value{
Returns a sp::SpatialPointsDataFrame of cross section lines converted
to points.
}
\description{
Converts a `fluvgeo` cross section data structure from lines
to points.
}
\examples{
# Retrieve a cross section data structure
xs_sp <- fluvgeo::sin_riffle_floodplain_dims_L3_sp

# Call the xs2pts function for a cross section
xs_pts <- xs2pts(xs_sp)

}
