% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_bearing.R
\name{xs_bearing}
\alias{xs_bearing}
\title{Cross Section Bearing}
\usage{
xs_bearing(cross_section)
}
\arguments{
\item{cross_section}{sf; A cross section lines feature class.}
}
\value{
cross_section sf object with a new field bearing
}
\description{
Calculate the flow direction bearing of the cross section.
}
