% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_compare_plot_L2.R
\name{xs_compare_plot_L2}
\alias{xs_compare_plot_L2}
\title{Overplot multiple cross section profile surveys}
\usage{
xs_compare_plot_L2(
  stream,
  xs_number,
  xs_pts_sf_list,
  bankfull_elevation,
  aspect_ratio = 0.5
)
}
\arguments{
\item{stream}{character; The name of the stream.}

\item{xs_number}{integer; The cross section `Seq` number.}

\item{xs_pts_sf_list}{list; a list of `sf` objects of cross
section points, one for each survey time period
to be graphed. Survey list items must be tagged
with the survey label to be used in the graph
legend.}

\item{bankfull_elevation}{numeric; The detrended bankfull elevation (in
feet) that is used to calculate hydraulic
geometry.}

\item{aspect_ratio}{numeric; The aspect ratio of the graph.}
}
\value{
A ggplot2 object.
}
\description{
Produces a cross section profile plot for the specified cross
section for the input surveys.
}
\details{
This function is used to plot the cross section profile from a
series of \code{xs_points} data frames representing multiple surveys.
}
\seealso{
The \code{xs_compre_plot} function requires a \code{xs_points}
dataframe. See the \code{sin_xs_points} package dataset for an example of
this format of cross section data produced by the \code{FluvialGeomorph}
ArcGIS toolbox.
}
