% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_dimensions.R
\name{xs_dimensions}
\alias{xs_dimensions}
\title{Calculate cross section and regional metrics for a set of streams,
    regions, and detrended bankfull elevations}
\usage{
xs_dimensions(xs_points, streams, regions, bankfull_elevations)
}
\arguments{
\item{xs_points}{data frame; a data frame of cross section points}

\item{streams}{character vector; The stream names in the study
area.}

\item{regions}{character; The regions that a dimension will be
calculated for. See the regional_curves$region
field for a complete list.}

\item{bankfull_elevations}{numeric vector; The bankfull elevations (units:
detrended feet) that are used to calculate
hydraulic geometry.}
}
\value{
A data frame of cross section hydraulic geometry dimensions (area,
   width, depth, discharge) by cross section, xs type, and bankfull
   elevation.
   \describe{
       \item{reach_name}{character; The name of the stream.}
       \item{cross_section}{numeric; The cross section unique identifier.
                        Seq is only unique within a reach.}
       \item{xs_type}{character; A string indicating how the cross section
                      was derived. "DEM derived cross section" denotes
                      dimensions calculated from the DEM and "<Region
                      Name>" denotes that the dimensions were calculated
                      from that regions regional curve.}
       \item{bankfull_elevation}{numeric; The detrended bankfull elevation
                      (in feet) that is used to calculate hydraulic
                      geometry.}
       \item{drainage_area}{numeric; The area of the watershed upstream
                      from this cross section, units: square miles.}
       \item{xs_area}{numeric; The cross sectional area at the specified
                      detrended bankfull elevation, units: square feet.}
       \item{xs_width}{numeric; The cross section width at the specified
                      detrended bankfull elevation, units: feet.}
       \item{xs_depth}{numeric; The maximum depth at the specified
                      detrended bankfull elevation, units: detrended
                      feet.}
       \item{discharge}{numeric; The estimated discharge at the specified
                        drainage area.}
   }
}
\description{
Calculates a data frame of hydraulic geometry dimensions
    (area, width, depth) for a set of streams, regions, and detrended bankfull elevationsby cross section, xs type, and bankfull elevation.
}
\examples{
# Extract attribute data from the fluvgeo::sin_xs_points SpatialPointsDataFrame
sin_xs_points_df <- fluvgeo::sin_riffle_channel_points_sp@data

# Call the xs_dimensions function with test data
streams <- c("Sinsinawa")
regions <- c("Eastern United States", "IN Central Till Plain")
bankfull_elevations = seq(103, 104, 0.1)

sin <- xs_dimensions(xs_points = sin_xs_points_df,
                     streams = streams,
                     regions = regions,
                     bankfull_elevations = bankfull_elevations)

}
\seealso{
The \code{xs_dimensions} function calls the
\code{\link{xs_regional_metrics}} function, which calls the
\code{\link{xs_metrics}} function, which calls the
\code{\link{xs_geometry}} function.
}
