% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_geometry.R
\name{xs_geometry}
\alias{xs_geometry}
\title{Calculate hydraulic geometry for a single cross section.}
\usage{
xs_geometry(xs_points, bankfull_elevation)
}
\arguments{
\item{xs_points}{character; a data frame of cross section points.
Must be a single cross section.}

\item{bankfull_elevation}{numeric; The detrended bankfull elevation (in
feet) that is used to calculate hydraulic
geometry.}
}
\value{
A data frame of hydraulic dimensions at the specified detrended
   bankfull elevation. The data frame contains the fields:
   \describe{
       \item{xs_width}{The cross section width at the specified detrended
                       bankfull elevation}
       \item{xs_depth}{The maximum depth at the specified detrended bankfull
                       elevation}
       \item{xs_area}{The cross sectional area at the specified detrended
                      bankfull elevation}
       \item{bankfull_elev}{The ground elevation of the detrended bankfull
                            elevation}
   }
}
\description{
Calculates the hydraulic geometry for the input cross
    section.
}
\note{
The cross section points used as input to this function must represent
      only one cross section. See the documentation for the example
      \code{sin_xs_points} data frame for the specification of this input
      \code{sp::SpatialPointsDataFrame} object.
}
\examples{
# Extract the attribute data from the sin_xs_points SpatialPointsDataFrame
#     object
sin_xs_points_df <- sin_xs_points@data

# Subset the sin_xs_points data frame to contain only one cross section
#     (Seq = 4).
sin_xs_points_4 <- sin_xs_points_df[sin_xs_points_df$Seq == 4,]

# Calculate hydraulic geometry for a single cross section
xs_geometry(xs_points = sin_xs_points_4, bankfull_elevation = 103.5)

}
