% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_map.R
\name{xs_map}
\alias{xs_map}
\title{Cross section location map}
\usage{
xs_map(xs_fc, xs_number, dem, banklines, extent_factor = 7)
}
\arguments{
\item{xs_fc}{SpatialLinesDataFrame; a cross section lines
feature class}

\item{xs_number}{integer; The cross section identifier of the
requested cross section.}

\item{dem}{character; path to a dem raster}

\item{banklines}{SpatialLinesDataFrame; a banklines feature class}

\item{extent_factor}{numeric; A numeric value used to expand the map
extent around each cross section.}
}
\value{
a tmap object
}
\description{
Produces a cross section location map for the specified cross
section.
}
