% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_metric_ratios.R
\name{xs_metric_ratios}
\alias{xs_metric_ratios}
\title{Calculate metric ratios for each cross section}
\usage{
xs_metric_ratios(xs_dims)
}
\arguments{
\item{xs_dims}{data frame; A data frame of cross section
dimensions.}
}
\value{
Returns the input xs_dims data frame of cross sections with the
  calculated metric ratio variables added.
}
\description{
Calculates metric ratios for each cross section in the input
  xs_dims data frame.
}
