% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_metrics.R
\name{xs_metrics}
\alias{xs_metrics}
\title{Calculate cross section hydraulic geometry metrics}
\usage{
xs_metrics(xs_points, stream, xs_number, bankfull_elevation)
}
\arguments{
\item{xs_points}{character; a data frame of cross section points}

\item{stream}{character; The name of the stream.}

\item{xs_number}{integer; The cross section identifier of the
requested cross section.}

\item{bankfull_elevation}{numeric; The detrended bankfull elevation (in
feet) that is used to calculate hydraulic
geometry.}
}
\value{
A data frame of hydraulic dimensions for the specified cross section
   at the specified detrended bankfull elevation.
   \describe{
       \item{reach_name}{character; The name of the stream.}
       \item{cross_section}{numeric; The cross section unique identifier.
                        Seq is only unique within a reach.}
       \item{xs_type}{character; A string indicating how the cross section
                      was derived. "DEM derived cross section" denotes
                      dimensions calculated from the DEM and "<Region
                      Name>" denotes that the dimensions were calculated
                      from that regions regional curve.}
       \item{bankfull_elevation}{numeric; The detrended bankfull elevation
                      (in feet) that is used to calculate hydraulic
                      geometry.}
       \item{drainage_area}{numeric; The area of the watershed upstream
                      from this cross section in square miles.}
       \item{xs_area}{numeric; The cross sectional area at the specified
                      detrended bankfull elevation.}
       \item{xs_width}{numeric; The cross section width at the specified
                       detrended bankfull elevation.}
       \item{xs_depth}{numeric; The maximum depth at the specified
                       detrended bankfull elevation.}
       \item{fp_area}{numeric; The cross sectional area at the flood prone
                      elevation.}
       \item{fp_width}{numeric; The cross section width at the flood prone
                       elevation.}
       \item{fp_depth}{numeric; The maximum depth at the flood prone
                       elevation.}
       \item{xs_width_depth_ratio}{numeric; The bankfull width to bankfull
                       depth ratio.}
       \item{xs_entrenchment_ratio}{numeric; The entrenchment ratio (flood
                       prone width) / (bankfull width).}
       \item{watersurface_elev}{numeric; The water surface elevation.}
       \item{bankfull_elev}{numeric; The bankfull elevation.}
       \item{floodprone_elev}{numeric; The flood prone elevation.}
   }
}
\description{
Calculates the hydraulic geometry metrics bankfull (width,
    depth, area, elevation), floodprone (width, depth, area, elevation),
    width-depth ratio, entrenchment ratio, and water surface elevation for
    the input cross section.
}
\details{
Add methods description for calculating flood prone, width-depth
    ratio, and entrenchment ratio. Calls \code{xs_geometry()}.
}
