% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_metrics_plot.R
\name{xs_metrics_plot}
\alias{xs_metrics_plot}
\title{Plot cross section metrics}
\usage{
xs_metrics_plot(reach_xs_dims, features_sp, label_xs = TRUE)
}
\arguments{
\item{reach_xs_dims}{data frame; a data frame of cross section
dimensions.}

\item{features_sp}{SpatialPointsDataFrame of infrastructure features}

\item{label_xs}{boolean; Draw the cross section locations?}
}
\value{
A ggplot2 object.
}
\description{
Produces a longitudinal plot of cross section metrics for the
input stream reach.
}
\examples{
# Extract cross section dimension data
sin_xs_dims_df <- fluvgeo::sin_riffle_floodplain_dims_planform_sp@data

# Call the xs_metrics_plot function
sin_metrics <- xs_metrics_plot(reach_xs_dims = sin_xs_dims_df)

# Print the graph
print(sin_metrics)

}
\seealso{
The \code{xs_metrics_plot} function requires a \code{xs_dimensions}
SpatialPointsDataFrame. See the \code{sin_xs_dimensions} package dataset for
an example of this format of cross section data produced by the
\code{FluvialGeomorph} ArcGIS toolbox.
}
