% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_metrics_plot2.R
\name{xs_metrics_plot2}
\alias{xs_metrics_plot2}
\title{Plot cross section metrics}
\usage{
xs_metrics_plot2(reach_xs_dims, label_xs = TRUE)
}
\arguments{
\item{reach_xs_dims}{data frame; a data frame of cross section
dimensions.}

\item{label_xs}{boolean; Draw the cross section locations?}
}
\value{
A ggplot2 object.
}
\description{
Produces a longitudinal plot of cross section metrics for the
input stream reach.
}
\examples{
# Extract data from the fgm::sin_xs_dimensions SpatialPointsDataFrame
sin_xs_dims_df <- fgm::sin_xs_dimensions@data

# Call the xs_plot function
sin_profile <- xs_metrics_plot(reach_xs_dims = sin_xs_dims_df)

# Print the graph
sin_profile

}
\seealso{
The \code{xs_metrics_plot} function requires a \code{xs_dimensions}
dataframe. See the \code{sin_xs_dimensions} package dataset for an
example of this format of cross section data produced by the
\code{FluvialGeomorph} ArcGIS toolbox.
}
