% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_metrics_plot_L1.R
\name{xs_metrics_plot_L1}
\alias{xs_metrics_plot_L1}
\title{Plot Level 1 Cross Section Metrics}
\usage{
xs_metrics_plot_L1(
  xs_dims_sf,
  features_sf,
  label_xs = TRUE,
  xs_label_freq = 10,
  profile_units = "kilometers"
)
}
\arguments{
\item{xs_dims_sf}{SimpleFeatures data frame of Level 1 cross section
dimensions.}

\item{features_sf}{SimpleFeatures data frame of infrastructure features.}

\item{label_xs}{logical; Draw the cross section locations?}

\item{xs_label_freq}{numeric; An integer indicating the frequency of
cross section labels.}

\item{profile_units}{character; the units of the longitudinal profile.
One of "kilometers", "meters", "miles", or "feet"}
}
\value{
A ggplot2 object.
}
\description{
Produces a longitudinal plot of Level 1 cross section metrics
for the input stream reach.
}
