% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_plot.R
\name{xs_plot}
\alias{xs_plot}
\title{Plot cross section profile}
\usage{
xs_plot(xs_points, stream, xs_number, bankfull_elevation)
}
\arguments{
\item{xs_points}{character; a data frame of cross section points}

\item{stream}{character; The name of the stream.}

\item{xs_number}{integer; The cross section identifier of the
requested cross section.}

\item{bankfull_elevation}{numeric; The detrended bankfull elevation (in
feet) that is used to calculate hydraulic
geometry.}
}
\value{
A ggplot2 object.
}
\description{
Produces a cross section profile plot for the specified cross
section and displayes the specified bankfull elevation.
}
\details{
This function is used to plot the cross section profile from a
\code{xs_points} data frame.
}
\examples{
# Extract attribute data from the fgm::sin_xs_points SpatialPointsDataFrame
sin_xs_points_df <- fgm::sin_xs_points@data

# Call the xs_plot function
sin_4_plot <- xs_plot(xs_points = sin_xs_points_df,
                      stream = "Sinsinawa",
                      xs_number = 4,
                      bankfull_elevation = 103.5)
sin_4_plot

}
