% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_profile_plot.R
\name{xs_profile_plot}
\alias{xs_profile_plot}
\title{Plot stream longitudinal profile}
\usage{
xs_profile_plot(reach_xs_dims, features = NULL, label_xs = TRUE)
}
\arguments{
\item{reach_xs_dims}{data frame; a data frame of cross section
dimensions.}

\item{features}{data frame; a data frame of river features}

\item{label_xs}{boolean; Draw the cross section locations?}
}
\value{
A ggplot2 object.
}
\description{
Produces a longitudinal profile plot for the cross sections
of the input stream reach.
}
\examples{
# Extract data from the fgm::sin_xs_dimensions SpatialPointsDataFrame
sin_xs_dims_df <- fgm::sin_xs_dimensions@data

# Call the xs_plot function
sin_profile <- xs_profile_plot(reach_xs_dims = sin_xs_dims_df)

# Print the graph
sin_profile

}
\seealso{
The \code{profile_plot} function requires a \code{xs_dimensions}
dataframe. See the \code{sin_xs_dimensions} package dataset for an
example of this format of cross section data produced by the
\code{FluvialGeomorph} ArcGIS toolbox.
}
