% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_pts_classify.R
\name{xs_pts_classify}
\alias{xs_pts_classify}
\title{Classify Cross Section Points}
\usage{
xs_pts_classify(xs_pts, channel_poly, floodplain_poly, buffer_distance)
}
\arguments{
\item{xs_pts}{sf object; A cross_section_points object.}

\item{channel_poly}{sf object; A channel_poly object.}

\item{floodplain_poly}{sf object; A floodplain_poly object.}

\item{buffer_distance}{numeric; The buffer distance used to extend the
channel and floodplain polygons. Linear units of
sf objects.}
}
\value{
An sf object representing xs_pts whose points are classified as
         falling inside the floodplain or channel.
}
\description{
Classify cross section points that are located within the
             floodplain and channel.
}
