% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_regional_metrics.R
\name{xs_regional_metrics}
\alias{xs_regional_metrics}
\title{Calculate cross section and regional hydraulic geometry metrics}
\usage{
xs_regional_metrics(xs_points, stream, xs_number, bankfull_elevation,
  region)
}
\arguments{
\item{xs_points}{data frame; a data frame of cross section points}

\item{stream}{character; The name of the stream.}

\item{xs_number}{integer; The cross section identifier of the
requested cross section.}

\item{bankfull_elevation}{numeric; The detrended bankfull elevation
(units: NAVD88 feet) that is used to calculate
hydraulic geometry.}

\item{region}{character; The region that a dimension will be
calculated for. See the regional_curves$region
field for a complete list.}
}
\value{
A data frame of hydraulic dimensions for the specified cross
   section at the specified detrended bankfull elevation and the regional
   hydraulic dimensions.
   \describe{
       \item{reach_name}{character; The name of the stream.}
       \item{cross_section}{numeric; The cross section unique identifier.
                        Seq is only unique within a reach.}
       \item{xs_type}{character; A string indicating how the cross section
                      was derived. "DEM derived cross section" denotes
                      dimensions calculated from the DEM and "<Region
                      Name>" denotes that the dimensions were calculated
                      from that regions regional curve.}
       \item{bankfull_elevation}{numeric; The detrended bankfull elevation
                      (in feet) that is used to calculate hydraulic
                      geometry.}
       \item{drainage_area}{numeric; The area of the watershed upstream
                      from this cross section, units: square miles.}
       \item{xs_area}{numeric; The cross sectional area at the specified
                      detrended bankfull elevation, units: square feet.}
       \item{xs_width}{numeric; The cross section width at the specified
                       detrended bankfull elevation, units: feet.}
       \item{xs_depth}{numeric; The maximum depth at the specified
                       detrended bankfull elevation, units: detrended
                       feet.}
   }
}
\description{
Calculates cross section geometry and regional hydraulic
geometry dimensions for the specified cross section at the specified
bankfull elevation.
}
\examples{
# Extract attribute data from the fgm::sin_xs_points SpatialPointsDataFrame
sin_xs_points_df <- fgm::sin_xs_points@data

# Call the xs_metrics function
sin_4 <- xs_regional_metrics(xs_points = sin_xs_points_df,
                             stream = "Sinsinawa",
                             xs_number = 4,
                             bankfull_elevation = 103.5,
                             region = "Eastern United States")

}
\seealso{
The \code{xs_regional_metrics} function calls the \code{\link{xs_metrics}}
function which calls the \code{\link{xs_geometry}} function. The
\code{xs_regional_metrics} function is called by the
\code{\link{xs_dimensions}} function.
}
