% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_upstream.R
\name{xs_upstream}
\alias{xs_upstream}
\title{Calculate upstream cross section}
\usage{
xs_upstream(cross_section)
}
\arguments{
\item{cross_section}{sf line object, The cross section feature.}
}
\value{
sf line object
}
\description{
Calculates the upstream cross section coordinates, updates
cross section end points, reverses incorrectly digitized cross sections.
}
