% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_section.R
\name{check_cross_section}
\alias{check_cross_section}
\title{Check the validity of an `fgm` `cross_section` data structure}
\usage{
check_cross_section(cross_section, step = c("assign_ids",
  "watershed_area", "river_position", "station_points", "loop_bend"))
}
\arguments{
\item{cross_section}{SpatialLinesDataFrame: a `cross_section` data
structure used by the fgm package.}

\item{step}{character; last completed processing step. One of
"assign_ids", "watershed_area", "river_position",
"station_points", "loop_bend"}
}
\value{
Returns TRUE if the `cross_section` data structure matches the
requirements. The function throws an error for a data structure not matching
the data specification. Returns errors describing how the the data structure
doesn't match the requirement.
}
\description{
Checks that the input data structure `cross_section` meets
the requirements for this data structure.
}
\details{
Cross section feature classes evolve as different steps are
performed on them. The `step` parameter allows a `cross section` data
structure to be checked throughout its lifecycle. Each step defines a
changing set of requirements for the `cross section` data structure.
}
