% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_section_dimensions.R
\name{check_cross_section_dimensions}
\alias{check_cross_section_dimensions}
\title{Check the validity of an `fluvgeo` `cross_section_dimension` data
structure}
\usage{
check_cross_section_dimensions(
  cross_section_dimensions,
  step = c("cross_section_dimensions", "shear_stress", "stream_power", "planform",
    "metric_ratios")
)
}
\arguments{
\item{cross_section_dimensions}{a data frame or SpatialLinesDataFrame: a
`cross_section_dimension` data structure
used by the fluvgeo package.}

\item{step}{character; last completed processing step. One of
"cross_section_dimensions", "shear_stress,
"stream_power", "planform", "metric_ratios"}
}
\value{
Returns TRUE if the `cross_section_dimension` data structure matches
the requirements. The function throws an error for a data structure not
matching the data specification. Returns errors describing how the the data
structure doesn't match the requirement.
}
\description{
Checks that the input data structure `cross_section_dimension`
meets the requirements for this data structure.
}
\details{
Cross section dimension feature classes evolve as different steps
are performed on them. The `step` parameter allows a `cross section_dimension`
data structure to be checked throughout its lifecycle. Each step defines a
changing set of requirements for the `cross section_dimension` data structure.
}
\examples{
# Create testing data
## Step: cross_section_dimensions
xs_dims <- cross_section_dimensions(xs = fluvgeo::sin_riffle_channel_sp,
                                    xs_points = fluvgeo::sin_riffle_channel_points_sp,
                                    bankfull_elevation = 103,
                                    lead_n = 1,
                                    use_smoothing = TRUE,
                                    loess_span = 0.5,
                                    vert_units = "ft")
## Step: shear stress
xs_dims_ss <- shear_stress(xs_dims)

## Step: stream_power
xs_dims_sp <- stream_power(xs_dims_ss,
                           discharge_method = "regional_curve",
                           region = "Illinois River",
                           drainage_area = 41)

## Step: planform
xs_dims_plan <- planform_dimensions(fluvgeo::sin_riffle_floodplain_dims_sp,
                                    fluvgeo::sin_bankline_points_sp)

## Step: metric_ratios
xs_dims_ratios <- xs_metric_ratios(xs_dims_plan)

}
