% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{cross_section}
\alias{cross_section}
\title{Cross Section}
\usage{
cross_section(xs, flowline_points)
}
\arguments{
\item{xs}{sf object; A newly digitized set of cross sections.}

\item{flowline_points}{sf object; A flowline_points feature.}
}
\value{
an sf object representing a valid set of cross section line features
}
\description{
Creates a valid set of cross section features from newly 
             digitized stream cross section lines.
}
