% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri_raster2RasterLayer.R
\name{esri_raster2RasterLayer}
\alias{esri_raster2RasterLayer}
\title{Convert an ESRI raster to a RasterLayer}
\usage{
esri_raster2RasterLayer(raster_path)
}
\arguments{
\item{raster_path}{character; A path to an ESRI raster.}
}
\value{
A `raster::RasterLayer` object.
}
\description{
Converts an ESRI raster to a raster::RasterLayer.
}
\details{
Rasters stored in ESRI Geodatabases can currently only be read using
the `arcgisbinding`
}
