% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bankfull.R
\name{estimate_bankfull}
\alias{estimate_bankfull}
\title{Estimate bankfull report}
\usage{
estimate_bankfull(
  stream,
  xs_dims_fc,
  xs_points_fc,
  features_fc,
  regions,
  bankfull_elevations,
  bf_estimate,
  stat,
  label_xs,
  profile_units,
  output_dir,
  output_format
)
}
\arguments{
\item{stream}{character; The stream name. The stream name must
match a stream name in `ReachName` field in the
other parameters.}

\item{xs_dims_fc}{character; The path to the "xs_dims" feature
class.}

\item{xs_points_fc}{character; The path to a `xs_points` feature
class for the first time period.}

\item{features_fc}{character; The path to a `features` feature class.}

\item{regions}{character; The regions that a dimension will be
calculated for. See the regional_curves$region
field for a complete list.}

\item{bankfull_elevations}{numeric vector; The bankfull elevations (units:
detrended feet) that are used to calculate
hydraulic geometry.}

\item{bf_estimate}{numeric; The detrended bankfull elevation (in
feet) that is used to calculate hydraulic
geometry.}

\item{stat}{character; The statistic to graph "RMSE", "MAE"
(the default).}

\item{label_xs}{logical; Label cross sections?}

\item{profile_units}{character; the units of the longitudinal profile.
One of "kilometers", "meters", "miles", or "feet"}

\item{output_dir}{character; The output directory.}

\item{output_format}{character; The output format of the report. One
of "html_document", "word_document",
"pdf_document".}
}
\value{
Produces a FluvialGeomorph Estimate Bankfull report in the
`output_dir` in the requested file format.
}
\description{
Produces the Estimate Bankfull report.
}
\examples{


}
