% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bankfull.R
\name{estimate_bankfull}
\alias{estimate_bankfull}
\title{Estimate Bankfull Report}
\usage{
estimate_bankfull(
  stream,
  flowline_fc,
  xs_dims_fc,
  xs_points_ch_1,
  xs_points_ch_2,
  xs_points_ch_3,
  xs_points_ch_4,
  xs_points_fp_1,
  xs_points_fp_2,
  xs_points_fp_3,
  xs_points_fp_4,
  survey_name_1,
  survey_name_2,
  survey_name_3,
  survey_name_4,
  features_fc,
  dem,
  show_xs_map,
  regions,
  bankfull_elevations,
  bf_estimate,
  stat,
  label_xs,
  profile_units,
  aerial,
  elevation,
  xs_label_freq,
  exaggeration,
  extent_factor,
  output_dir,
  output_format
)
}
\arguments{
\item{stream}{character; The stream name. The stream name must
match a stream name in `ReachName` field in the
other parameters.}

\item{flowline_fc}{character; The path to the `flowline` feature
class.}

\item{xs_dims_fc}{character; The path to the
`riffle_channel_dims_L2` feature class.
This is for the "base year" survey.}

\item{xs_points_ch_1}{character; The path to the `riffle_channel_points`
feature class for the "base year".}

\item{xs_points_ch_2}{character; The path to the `riffle_channel_points`
feature class for the second time period.}

\item{xs_points_ch_3}{character; The path to the `riffle_channel_points`
feature class for the third time period.}

\item{xs_points_ch_4}{character; The path to the `riffle_channel_points`
feature class for the fourth time period.}

\item{xs_points_fp_1}{character; The path to the
`riffle_floodplain_points` feature class for
the "base year".}

\item{xs_points_fp_2}{character; The path to the
`riffle_floodplain_points` feature class for the
second time period.}

\item{xs_points_fp_3}{character; The path to the
`riffle_floodplain_points` feature class for the
third time period.}

\item{xs_points_fp_4}{character; The path to the
`riffle_floodplain_points` feature class for the
fourth time period.}

\item{survey_name_1}{character: The name or date of the "base year"
survey.}

\item{survey_name_2}{character: The name or date of the second survey.}

\item{survey_name_3}{character: The name or date of the third survey.}

\item{survey_name_4}{character: The name or date of the fourth survey.}

\item{features_fc}{character; The path to a `features` feature class.}

\item{dem}{character; The path to the DEM raster.}

\item{show_xs_map}{logical; Add the cross section maps to the report?}

\item{regions}{character; The regions that a dimension will be
calculated for. See the regional_curves$region
field for a complete list.}

\item{bankfull_elevations}{numeric vector; The bankfull elevations (units:
detrended feet) that are used to calculate
hydraulic geometry.}

\item{bf_estimate}{numeric; The detrended bankfull elevation (in
feet) that is used to calculate hydraulic
geometry.}

\item{stat}{character; The statistic to graph "RMSE", "MAE"
(the default).}

\item{label_xs}{logical; Label cross sections?}

\item{profile_units}{character; The units of the longitudinal profile.
One of "kilometers", "meters", "miles", or "feet"}

\item{aerial}{logical; Display an overview map with an aerial
photo background?}

\item{elevation}{logical; Display an overview map with an elevation
background?}

\item{xs_label_freq}{numeric; An integer indicating the frequency of
cross section labels.}

\item{exaggeration}{numeric; The degree of terrain exaggeration.}

\item{extent_factor}{numeric; The amount the extent is expanded around
the cross section feature class. Values greater
than one zoom out, values less than one zoom in.}

\item{output_dir}{character; The output directory.}

\item{output_format}{character; The output format of the report. One
of "html_document", "word_document",
"pdf_document".}
}
\value{
Produces a FluvialGeomorph Estimate Bankfull report in the
`output_dir` in the requested file format.
}
\description{
Produces the FluvialGeomorph Estimate Bankfull report.
}
