% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc2sf.R
\name{fc2sf}
\alias{fc2sf}
\title{Convert ESRI .gdb feature class to `sf`}
\usage{
fc2sf(fc_path)
}
\arguments{
\item{fc_path}{character; Path to the feature class (e.g.,
"C:/folder/geodatabase.gdb/feature_dataset/feature_class").}
}
\value{
An `sf` object.
}
\description{
Converts an ESRI geodatabase (.gdb) feature class to the `sf`
spatial object format.
}
