% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowline_metrics.R
\name{flowline_metrics}
\alias{flowline_metrics}
\title{Level 1 longitudinal metrics}
\usage{
flowline_metrics(gradient_sp, features_sp)
}
\arguments{
\item{gradient_sp}{SpatialPointsDataFrame of gradient points.}

\item{features_sp}{SpatialPointsDataFrame of infrastructure features}
}
\value{
A ggplot2 object.
}
\description{
Produces a longitudinal plot of Level 1 metrics (water surface
elevation, slope, and sinuosity) for the input stream reach.
}
