% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowline_points.R
\name{flowline_points}
\alias{flowline_points}
\title{Flowline Points}
\usage{
flowline_points(flowline, dem, station_distance)
}
\arguments{
\item{flowline}{sf object; A flowline object.}

\item{dem}{terra SpatRast object; A DEM.}

\item{station_distance}{numeric; Disired distance between the 
flowline_points (units: meters)}
}
\value{
an sf object of type point
}
\description{
Converts a stream flowline to a route using the km to mouth 
parameter and creates an sf object stream profile points.
}
