% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{fluvgeo}
\alias{fluvgeo-package}
\alias{fluvgeo}
\title{A Package for Fluvial Geomorphic Analysis}
\description{
This package contains a wide range of functions for performing fluvial geomorphic analysis. This package is designed for use with the FluvialGeomorph Python package. The FluvialGeomorph package contains an ArcGIS toolbox that can be used to extract spatial data from high- resolution terrain data. FluvialGeomorph uses functions in this package to perform a wide range of fluvial geomorphic analysis (e.g., extract stream channel dimensions, help choose bankfull elevation for ungaged streams).
}
\details{
Place some other package details here. And then some more.
}
\section{Installation}{

The datasets used as input into this package can be created using the
FluvialGeomorph-toolbox
\url{https://github.com/FluvialGeomorph/FluvialGeomorph-toolbox} Python
package. From within the GIS Windows application \code{ArcGIS Pro}, the
\code{arcgisbinding} R package can be installed by following the
instructions at
\url{https://r-arcgis.github.io/assets/arcgisbinding-vignette.html}.
No other functions within this package require ESRI \code{ArcGIS Pro}.
}

\references{
\describe{
  \item{{ArcGIS Pro}\url{https://pro.arcgis.com/}}
  \item{\code{arcgisbinding}}
  \url{https://r-arcgis.github.io/assets/arcgisbinding-vignette.html}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FluvialGeomorph/fluvgeo}
  \item \url{https://fluvialgeomorph.github.io/fluvgeo/}
  \item Report bugs at \url{https://github.com/FluvialGeomorph/fluvgeo/issues}
}

}
\author{
\strong{Maintainer}: Barrie Chileen Martinez \email{Barrie.V.Chileen@usace.army.mil} (\href{https://orcid.org/0000-0002-6960-8167}{ORCID})

Authors:
\itemize{
  \item Michael Dougherty \email{Michael.P.Dougherty@usace.army.mil} (\href{https://orcid.org/0000-0002-1465-5927}{ORCID})
  \item Christopher Haring \email{Christopher.P.Haring@usace.army.mil}
}

Other contributors:
\itemize{
  \item Thomas Darby [reviewer]
  \item Davi Michl [reviewer]
}

}
\keyword{internal}
