% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_leaflet.R
\name{get_results_leaflet}
\alias{get_results_leaflet}
\title{Get Results Leaflet}
\usage{
get_results_leaflet(fl, xs, dem)
}
\arguments{
\item{fl}{sf object;}

\item{xs}{sf object;}

\item{dem}{terra SpatRast object;}
}
\value{
a leaflet object
}
\description{
Get Results Leaflet
}
