% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reach_overview.R
\name{map_reach_overview}
\alias{map_reach_overview}
\title{Reach Overview Map}
\usage{
map_reach_overview(flowline, cross_section)
}
\arguments{
\item{flowline}{SpatialLinesDataFrame; a flowline feature class}

\item{cross_section}{SpatialLinesDataFrame; a cross section feature
class}
}
\value{
a tmap object
}
\description{
Produces a reach overview map displaying cross section
locations over an aerial image.
}
\examples{
# Use the fluvgeo::sin_flowline_sp SpatialLinesDataFrame
sin_flowline_sp <- fluvgeo::sin_flowline_sp

# Use the fluvgeo::sin_riffle_floodplain_sp SpatialLinesDataFrame
sin_riffle_channel_sp <- fluvgeo::sin_riffle_channel_sp

# Create the map
sin_map <- map_reach_overview(sin_flowline_sp, sin_riffle_channel_sp)

# Print the map
print(sin_map)

}
