% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reach_overview.R
\name{map_reach_overview}
\alias{map_reach_overview}
\title{Reach Overview Map}
\usage{
map_reach_overview(
  flowline_sf,
  cross_section_sf,
  background = "none",
  xs_label_freq = 1,
  exaggeration = 20,
  extent_factor = 1.1
)
}
\arguments{
\item{flowline_sf}{sf object; A flowline feature class}

\item{cross_section_sf}{sf object; A cross section feature class}

\item{background}{character; The type of map background. One of
"aerial", "elevation", or "none".}

\item{xs_label_freq}{numeric; An integer indicating the frequency of
cross section labels.}

\item{exaggeration}{numeric; The degree of terrain exaggeration.}

\item{extent_factor}{numeric; The amount the extent is expanded around
the cross section feature class. Values greater
than one zoom out, values less than one zoom in.}
}
\value{
a tmap object
}
\description{
Produces a reach overview map displaying cross section
locations over an aerial image or elevation multi-direction shaded relief.
}
