% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_xs.R
\name{map_xs}
\alias{map_xs}
\title{Cross section location map}
\usage{
map_xs(cross_section, xs_number, dem, banklines, extent_factor = 7)
}
\arguments{
\item{cross_section}{SpatialLinesDataFrame; a cross section lines
feature class}

\item{xs_number}{integer; The cross section identifier of the
requested cross section.}

\item{dem}{character; path to a dem raster}

\item{banklines}{SpatialLinesDataFrame; a banklines feature class}

\item{extent_factor}{numeric; A numeric value used to expand the map
extent around each cross section.}
}
\value{
a tmap object
}
\description{
Produces a cross section location map for the specified cross
section.
}
