% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meander_width.R
\name{meander_width}
\alias{meander_width}
\title{Calculate the meander width for each loop}
\usage{
meander_width(bankline_points)
}
\arguments{
\item{bankline_points}{sf data.frame; a fluvgeo bankline_points data
structure}
}
\value{
Returns a data frame of loops with the calculated meander width in
units feet.
}
\description{
Calculates the meander width (channel belt width) for each loop
in the input `bankline_points` data frame.
}
\examples{
bankline_points <- fluvgeo::sin_bankline_points_sf
mw <- meander_width(bankline_points = bankline_points)

}
