% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planform.R
\name{planform}
\alias{planform}
\title{Calculate planform dimensions}
\usage{
planform(bankline_points)
}
\arguments{
\item{bankline_points}{SpatialPointsDataFrame; a `bankline_points` data
structure used by the fgm package.}
}
\value{
Returns a data frame of bends with the calculated planform
dimensions.
}
\description{
Calculates planform dimensions (bend radius of curvature,
meander wavelength, and meander belt width) for each bend and loop in the
input `bankline_points` data frame.
}
\examples{
pf <- planform(fgm::sin_bankline_points_sp)

}
