% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_sinuosity.R
\name{slope_sinuosity}
\alias{slope_sinuosity}
\title{Calculate slope and sinuosity}
\usage{
slope_sinuosity(
  channel_features,
  lead_n,
  lag_n,
  use_smoothing = TRUE,
  loess_span = 0.05,
  vert_units
)
}
\arguments{
\item{channel_features}{Spatial*DataFrame; an `fluvgeo` data structure of
channel features (i.e., cross section, flowline
points, etc.) Must have the following fields:
`ReachName`, `POINT_X`, `POINT_Y`, `POINT_M`, `Z`}

\item{lead_n}{numeric; The number of features to lead (upstream)
to calculate the slope and sinuosity. Must be an
integer.}

\item{lag_n}{numeric; The number of features to lag (downstream)
to calculate the slope and sinuosity. Must be an
integer.}

\item{use_smoothing}{boolean; determines if smoothed elevation values
are used to calculate gradient. values are:
TRUE, FALSE (default)}

\item{loess_span}{numeric; the loess regression span parameter,
defaults to 0.05}

\item{vert_units}{character; The vertical units. One of: "m" (meter),
"ft" (foot), "us-ft" (us survey foot)}
}
\value{
A dataframe of slope and sinuosity dimensions representing the
position of each feature within the channel.
}
\description{
Calculates the slope and sinuosity of the input channel
features.
}
\examples{
# Extract attribute data from the fluvgeo::sin_flowline_points_sp
# SpatialPointsDataFrame
sin_flowline_points_df <- fluvgeo::sin_flowline_points_sp@data

# Extract data from the fluvgeo::sin_riffle_channel_sp SpatialLinesDataFrame
sin_rifflechannel_df <- fluvgeo::sin_riffle_channel_sp@data

# Call the slope_sinuosity function for a flowline
sin_flowline_ss <- slope_sinuosity(sin_flowline_points_df,
                                   lead_n = 1000, lag_n = 0)

# Call the slope_sinuosity function for a cross section
sin_riffle_channel_ss <- slope_sinuosity(sin_riffle_channel_df,
                                 lead_n = 1, lag_n = 0,
                                 loess_span = 5)

}
