% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_sinuosity.R
\name{slope_sinuosity}
\alias{slope_sinuosity}
\title{Calculate slope and sinuosity}
\usage{
slope_sinuosity(channel_features, lead_lag, use_smoothing = TRUE,
  loess_span = 0.05)
}
\arguments{
\item{channel_features}{data frame; a data frame of channel features (i.e.,
cross section, flowline points, etc.)}

\item{lead_lag}{numeric; The number of features to lead/lag on
either side of each feature that will be used to
calculate the slope and sinuosity. Must be an
integer.}

\item{use_smoothing}{boolean; determines if smoothed elevation values
are used to calculate gradient. values are:
TRUE, FALSE (default)}

\item{loess_span}{numeric; the loess regression span parameter,
defaults to 0.05}
}
\value{
A dataframe of dimensions representing the position of each feature
within the channel.
}
\description{
Calculates the slope and sinuosity of the input channel
features.
}
\examples{
# Extract data from the fgm::sin_flowline_points SpatialPointsDataFrame
sin_flowline_points_df <- fgm::sin_flowline_points@data

# Call the slope_sinuosity function
sin_flowline_ss <- slope_sinuosity(sin_flowline_points_df,
                                   lead_lag = 1000)

}
