% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp2arc.R
\name{sp2arc}
\alias{sp2arc}
\title{Converts an \code{sp} object to an ArcGIS feature class}
\usage{
sp2arc(sp_object, fc_path)
}
\arguments{
\item{sp_object}{\code{sp} object}

\item{fc_path}{string; Path to the ArcGIS feature class.}
}
\value{
Writes the \code{sp} object to an ArcGIS feature class specified
    by path.
}
\description{
Converts an \code{sp} object to an ArcGIS feature class in
    the specified geodatabase feature class.
}
