% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_xs_points.R
\name{update_xs_points}
\alias{update_xs_points}
\title{Update \code{xs_points}}
\usage{
update_xs_points(ReachName, xs_points_reach, xs_points)
}
\arguments{
\item{ReachName}{character; The reach name of the new xs_points data.}

\item{xs_points_reach}{sp object; The sp object containing the xs_points
for the specified reach.}

\item{xs_points}{sp object; The sp object containing the xs_points for
the study area.}
}
\value{
An updated SpatialPointsDataFrame object containing the terrain data
    for the specified reach.
}
\description{
Updates the \code{xs_points} SpatialPointsDataFrame object
    with terrain data for the specified reach.
}
