% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs2pts.R
\name{xs2pts}
\alias{xs2pts}
\title{XS lines to points}
\usage{
xs2pts(xs_sf)
}
\arguments{
\item{xs_sf}{sf data frame; a cross section data structure
created by the fluvgeo package.}
}
\value{
Returns a sf::sf data frame of cross section lines converted
to points.
}
\description{
Converts a `fluvgeo` cross section data structure from lines
to points.
}
\examples{
# Retrieve a cross section data structure
xs_sf <- fluvgeo::sin_riffle_floodplain_dims_L3_sf

# Call the xs2pts function for a cross section
xs_pts <- xs2pts(xs_sf)

}
