% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_geometry.R
\name{xs_geometry}
\alias{xs_geometry}
\title{Calculate cross section width, depth, area.}
\usage{
xs_geometry(xs_points, detrend_elevation)
}
\arguments{
\item{xs_points}{data frame; a data frame of cross section points.
Must be a single cross section.}

\item{detrend_elevation}{numeric; The detrended elevation used to
calculate hydraulic geometry, units: detrended
feet.}
}
\value{
A data frame of hydraulic dimensions at the specified detrended
   elevation. The data frame contains the fields:
   \describe{
       \item{xs_width}{numeric; The cross section width at the specified
                       detrended elevation, units: feet.}
       \item{xs_depth}{numeric; The maximum depth at the specified
                       detrended elevation, units: feet.}
       \item{xs_area}{numeric; The cross sectional area at the specified
                      detrended elevation, units: square feet.}
       \item{ground_elev}{numeric; The ground elevation of the detrended
                          elevation, units: NAVD88 feet.}
   }
}
\description{
Calculates the hydraulic geometry (width, depth, area, and
    ground elevation) for the input cross section at the specified
    detrended elevation.
}
\details{
The cross section points used as input to this function must
   represent only one cross section. See the documentation for the example
   \code{sin_xs_points} data frame for the specification of this input
   \code{sp::SpatialPointsDataFrame} object.
}
\examples{
# Extract attribute data from the fgm::sin_xs_points SpatialPointsDataFrame
sin_xs_points_df <- sin_xs_points@data

# Subset sin_xs_points to contain only one cross section (Seq = 4)
sin_xs_points_4 <- sin_xs_points_df[sin_xs_points_df$Seq == 4,]

# Calculate hydraulic geometry for a single cross section
xs_geometry(xs_points = sin_xs_points_4, detrend_elevation = 103.5)

}
\seealso{
The \code{xs_geometry} function is called by the \code{\link{xs_metrics}}
function, which is called by the \code{\link{xs_regional_metrics}}
function, which is called by the \code{\link{xs_dimensions}} function.
}
