% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_metric_plot.R
\name{xs_metric_plot}
\alias{xs_metric_plot}
\title{Plot cross section metrics}
\usage{
xs_metric_plot(
  metric,
  reach_xs_dims,
  label_xs = TRUE,
  xs_label_freq = 10,
  profile_units = "kilometers"
)
}
\arguments{
\item{metric}{FluvialGeomorphicMetric object; the fluvial geomorphic
metric to be mapped}

\item{reach_xs_dims}{data frame; a data frame of cross section
dimensions.}

\item{label_xs}{logical; Draw the cross section locations?}

\item{xs_label_freq}{numeric; An integer indicating the frequency of
cross section labels.}

\item{profile_units}{character; the units of the longitudinal profile.
One of "kilometers", "meters", "miles", or "feet"}
}
\value{
A ggplot2 object.
}
\description{
Produces a longitudinal plot of cross section metrics for the
input stream reach.
}
\examples{
# Get feature class test data
xs_planform_fc <- file.path(system.file("extdata", "testing_data.gdb",
                                        package = "fluvgeo"),
                            "riffle_floodplain_dims_L3")

# Convert feature classes to an sf objects
xs_planform_sf   <- fluvgeo::fc2sf(xs_planform_fc)

# Create the fluvgeo::FluvialGeomorphicMetric object
wdr <- new(Class = "FluvialGeomorphicMetric",
           metric = "Width Depth Ratio",
           definition = "bankfull width / bankfull depth",
           variable = "xs_width_depth_ratio",
           threshold_breaks = c(0, 10, 20, Inf),
           threshold_labels = c("Incised",
                                "Stable",
                                "Overwidened"),
           source = "Dunn & Leopold, 1978")

metric_plot <- fluvgeo::xs_metric_plot(metric = wdr,
                                       reach_xs_dims = xs_planform_sf,
                                       label_xs = TRUE,
                                       profile_units = "miles")
print(metric_plot)

}
\seealso{
The \code{xs_metrics_plot} function requires a \code{xs_dimensions}
dataframe. See the \code{sin_xs_dimensions} package dataset for an
example of this format of cross section data produced by the
\code{FluvialGeomorph} ArcGIS toolbox.
}
