% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_metric_ratios.R
\name{xs_metric_ratios}
\alias{xs_metric_ratios}
\title{Calculate metric ratios for each cross section}
\usage{
xs_metric_ratios(xs_dimensions)
}
\arguments{
\item{xs_dimensions}{SpatialLinesDataFrame; A `cross_section_dimensions`
fluvgeo data structure.}
}
\value{
Returns the input xs_dims data frame of cross sections with the
  calculated metric ratio variables added.
}
\description{
Calculates metric ratios for each cross section in the input
  xs_dims data frame.
}
\examples{
# Calculate cross section planform dimensions
xs_dims_plan <- planform_dimensions(fluvgeo::sin_riffle_floodplain_dims_sp,
                                    fluvgeo::sin_bankline_points_sp)

# Calculate cross section metric ratios
xs_dims_ratios <- xs_metric_ratios(xs_dims_plan)


}
