% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_metrics_report.R
\name{xs_metrics_report}
\alias{xs_metrics_report}
\title{Cross Section Metrics Report}
\usage{
xs_metrics_report(xs_points, xs_dims_planform, flowline, dem, banklines,
  extent_factor, streams, regions, features, label_xs, output_dir,
  output_format)
}
\arguments{
\item{xs_points}{data frame; a data frame of a cross section points
data struture}

\item{xs_dims_planform}{SpatialLinesDataFrame; a
SpatialLinesDataFrame of a cross section planform
dimensions data structure}

\item{flowline}{SpatialLinesDataFrame; a SpatialLinesDataFrame of
a flowline data structure}

\item{dem}{character; path to a dem raster}

\item{banklines}{SpatialLinesDataFrame; a banklines feature class}

\item{extent_factor}{numeric; A numeric value used to expand the map
extent around each cross section}

\item{streams}{character vector; The stream names in the study
area.}

\item{regions}{character; The regions that a dimension will be
calculated for. See the regional_curves$region
field for a complete list.}

\item{features}{SpatialPointsDataFrame; a
SpatialPointsDataFrame of river features}

\item{label_xs}{boolean; Draw the cross section labels?}

\item{output_dir}{character; The output directory.}

\item{output_format}{character; The output format of the report. One
of "html_document", "word_document",
"pdf_document".}
}
\value{
A report written to the file system in the output fromat requested.
}
\description{
Produces the cross section metrics report.
}
\examples{
# Extract attribute data from the fluvgeo::sin_xs_points SpatialPointsDataFrame
sin_xs_points_df <- fluvgeo::sin_xs_points@data
sin_xs_dims_planform_df <- fluvgeo::sin_xs_dims_planform@data

# Set variable values
streams <- c("Sinsinawa")
regions <- c("Eastern United States", "IN Central Till Plain")
output_dir <- Sys.getenv("HOME")
output_format <- "html_document"

# Call the  function with test data
xs_metrics_report(xs_points = sin_xs_points_df,
                  xs_dims_planform = sin_xs_dims_planform_df,
                  streams = streams,
                  regions = regions,
                  output_dir = output_dir,
                  output_format = output_format)

}
