% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_metrics_report.R
\name{xs_metrics_report}
\alias{xs_metrics_report}
\title{Cross Section Metrics Report}
\usage{
xs_metrics_report(
  xs_points,
  xs_dims_planform,
  flowline,
  dem,
  banklines,
  extent_factor,
  streams,
  regions,
  features,
  label_xs,
  output_dir,
  output_format
)
}
\arguments{
\item{xs_points}{data frame; a data frame of a cross section points
data struture}

\item{xs_dims_planform}{SpatialLinesDataFrame; a
SpatialLinesDataFrame of a cross section planform
dimensions data structure}

\item{flowline}{SpatialLinesDataFrame; a SpatialLinesDataFrame of
a flowline data structure}

\item{dem}{character; path to a dem raster}

\item{banklines}{SpatialLinesDataFrame; a banklines feature class}

\item{extent_factor}{numeric; A numeric value used to expand the map
extent around each cross section}

\item{streams}{character vector; The stream names in the study
area.}

\item{regions}{character; The regions that a dimension will be
calculated for. See the regional_curves$region
field for a complete list.}

\item{features}{SpatialPointsDataFrame; a
SpatialPointsDataFrame of river features}

\item{label_xs}{boolean; Draw the cross section labels?}

\item{output_dir}{character; The output directory.}

\item{output_format}{character; The output format of the report. One
of "html_document", "word_document",
"pdf_document".}
}
\value{
A report written to the file system in the output fromat requested.
}
\description{
Produces the cross section metrics report.
}
\examples{
# Extract attribute data from the fluvgeo::sin_xs_points SpatialPointsDataFrame
sin_xs_points_df <- fluvgeo::sin_riffle_floodplain_points_sp@data

# Get a cross section planform dimensions feature class
sin_xs_dims_planform <- fluvgeo::sin_riffle_floodplain_dims_planform_sp

# Set variable values
dem <- system.file("extdata", "dem_1m.tif", package = "fluvgeo")
extent_factor <- 2
streams <- c("Sinsinawa")
regions <- c("Eastern United States", "IN Central Till Plain")
output_dir <- Sys.getenv("HOME")
output_format <- "word_document"

# Create the xs Metrics Report
xs_metrics_report(xs_points = sin_xs_points_df,
                  xs_dims_planform = sin_xs_dims_planform,
                  flowline = fluvgeo::sin_flowline_sp,
                  dem = dem,
                  banklines = fluvgeo::sin_banklines_sp,
                  extent_factor = extent_factor,
                  streams = streams,
                  regions = regions,
                  features = fluvgeo::sin_features_sp,
                  label_xs = TRUE,
                  output_dir = output_dir,
                  output_format = output_format)

}
