% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_profile_plot.R
\name{xs_profile_plot}
\alias{xs_profile_plot}
\title{Plot stream longitudinal profile}
\usage{
xs_profile_plot(
  reach_xs_dims_sp,
  features_sp = NULL,
  label_xs = TRUE,
  xs_label_freq = 10,
  profile_units = "kilometers"
)
}
\arguments{
\item{reach_xs_dims_sp}{SpatialPointsDataFrame of cross section
dimensions.}

\item{features_sp}{SpatialPointsDataFrame of infrastructure features}

\item{label_xs}{logical; Draw the cross section labels?}

\item{xs_label_freq}{numeric; An integer indicating the frequency of
cross section labels.}

\item{profile_units}{character; the units of the longitudinal profile.
One of "kilometers", "meters", "miles", or "feet"}
}
\value{
A ggplot2 object.
}
\description{
Produces a longitudinal profile plot for the cross sections
of the input stream reach.
}
\examples{
# Create cross section profile plot
profile_plot <- xs_profile_plot(reach_xs_dims_sp = fluvgeo::sin_riffle_floodplain_dims_planform_sp,
                                features_sp = fluvgeo::sin_features_sp,
                                label_xs = TRUE)

# Print the plot
print(profile_plot)

}
\seealso{
The \code{profile_plot} function requires a \code{xs_dimensions}
dataframe. See the \code{sin_xs_dimensions} package dataset for an
example of this format of cross section data produced by the
\code{FluvialGeomorph} ArcGIS toolbox.
}
